% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{access_token}
\alias{access_token}
\title{Get user's current access token}
\usage{
access_token(
  origin = funky::config_val("origin"),
  email = funky::config_val("email", require = TRUE)
)
}
\arguments{
\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}
}
\value{
Access token as a character scalar. \code{NA_character_} if no access token exists for \code{origin} and \code{email}.
}
\description{
Retrieves the access token belonging to the user with the specified \code{email} from the corresponding package environment. An access token is only available if
the user has \link[=sign_in]{signed in} or \link[=refresh_sign_in]{refreshed their access token} during the current \R session.
}
\seealso{
Other access_token: 
\code{\link{decode_access_token}()},
\code{\link{is_access_token_expired}()},
\code{\link{store_access_token}()}
}
\concept{access_token}
\keyword{internal}
