% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{app_settings}
\alias{app_settings}
\title{List NocoDB app settings}
\usage{
app_settings(
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)
}
\arguments{
\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
The application settings of the specified NocoDB server as a \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with (meta)data about the application settings of a NocoDB server via its
\code{GET /api/v2/meta/nocodb/info} API endpoint.
}
\seealso{
Other functions to manage NocoDB's application settings:
\code{\link{update_app_settings}()}
}
\concept{app_settings}
