% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{assemble_url}
\alias{assemble_url}
\title{Assemble NocoDB URL}
\usage{
assemble_url(
  ...,
  .scheme = "https",
  .hostname = pal::pkg_config_val("hostname")
)
}
\arguments{
\item{...}{Optional path components added to the base URL.}

\item{.scheme}{\href{https://en.wikipedia.org/wiki/URL#Syntax}{Scheme} to use in the assembled URL. A character scalar.}

\item{.hostname}{\href{https://en.wikipedia.org/wiki/Hostname}{Hostname} to use in the assembled URL. A character scalar.}
}
\value{
A character scalar.
}
\description{
Assemble NocoDB URL
}
\examples{
\dontrun{
nocodb:::assemble_url("api/v2/meta/bases/")}
}
\concept{url_assembly}
\keyword{internal}
