% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{base}
\alias{base}
\title{Get NocoDB bases metadata}
\usage{
base(
  base_id = base_id(),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  auth_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)
}
\arguments{
\item{base_id}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{auth_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API authentication token}. A character scalar.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with metadata about the specified base on a NocoDB server from its
\href{https://meta-apis-v2.nocodb.com/#tag/Base/operation/base-read}{\code{GET /api/v2/meta/bases/{base_id}}} API endpoint.
}
\seealso{
Other functions to manage NocoDB bases:
\code{\link{base_id}()},
\code{\link{bases}()},
\code{\link{create_base}()},
\code{\link{update_base}()}
}
\concept{bases}
