% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{base_id}
\alias{base_id}
\title{Get NocoDB base ID}
\usage{
base_id(
  title = pal::pkg_config_val(key = "base_title", pkg = this_pkg),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  auth_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)
}
\arguments{
\item{title}{NocoDB base title. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{auth_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API authentication token}. A character scalar.}
}
\value{
A character scalar if a base titled \code{title} exists, otherwise a zero-length character vector.
}
\description{
Returns the identifier of the base with the specified title on a NocoDB server.
}
\seealso{
Other functions to manage NocoDB bases:
\code{\link{base}()},
\code{\link{bases}()},
\code{\link{create_base}()},
\code{\link{update_base}()}
}
\concept{bases}
