% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{base_ui_acl}
\alias{base_ui_acl}
\title{List NocoDB base visibility rules}
\usage{
base_ui_acl(
  id_base = base_id(hostname = hostname, email = email, password = password, api_token =
    api_token),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)
}
\arguments{
\item{id_base}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with metadata about \href{https://docs.nocodb.com/data-sources/actions-on-data-sources#ui-access-control}{visibility rules (UI ACL) for the different views} in the specified base on a NocoDB server from its
\href{https://meta-apis-v2.nocodb.com/#tag/Base/operation/base-model-visibility-list}{\code{GET /api/v2/meta/bases/{id_base}/visibility-rules}} API endpoint.
}
\seealso{
Other functions to manage NocoDB base visibility rules (UI ACL):
\code{\link{update_base_ui_acl}()}

Other functions to manage NocoDB table views:
\code{\link{tbl_view_id}()},
\code{\link{tbl_views}()},
\code{\link{update_base_ui_acl}()}

Other functions to manage NocoDB bases:
\code{\link{base}()},
\code{\link{base_id}()},
\code{\link{base_users}()},
\code{\link{bases}()},
\code{\link{create_base}()},
\code{\link{delete_base}()},
\code{\link{delete_base_user}()},
\code{\link{invite_base_user}()},
\code{\link{resend_base_user_invitation}()},
\code{\link{update_base}()},
\code{\link{update_base_ui_acl}()},
\code{\link{update_base_user}()}
}
\concept{bases}
\concept{ui_acl}
\concept{views}
