% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{create_base}
\alias{create_base}
\title{Create NocoDB base}
\usage{
create_base(
  title = pal::pkg_config_val(key = "base_title", pkg = this_pkg),
  description = NULL,
  color = "#36BFFF",
  show_null_and_empty_in_filter = TRUE,
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  auth_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)
}
\arguments{
\item{title}{Title of the new base. A character scalar.}

\item{description}{Description of the new base. A character scalar.}

\item{color}{Color to use for the base icon. A hexadecimal color code like "#36BFFF".}

\item{show_null_and_empty_in_filter}{Whether or not to distinguish \code{NULL}s from empty values in column filters. If \code{FALSE}, \code{NULL} and empty fields are
treated alike.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{auth_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API authentication token}. A character scalar.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} containing metadata about the newly created NocoDB base.
}
\description{
Adds a new base on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/Base/operation/base-create}{\code{POST /api/v2/meta/bases}} API endpoint.
}
\seealso{
Other functions to manage NocoDB bases:
\code{\link{base}()},
\code{\link{base_id}()},
\code{\link{bases}()},
\code{\link{update_base}()}
}
\concept{bases}
