% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{create_data_src}
\alias{create_data_src}
\title{Create NocoDB data source}
\usage{
create_data_src(
  alias = NULL,
  type = c("mssql", "mysql", "pg", "sqlite3"),
  config = NULL,
  inflection_column = c("none", "camelize"),
  inflection_table = c("none", "camelize"),
  enabled = TRUE,
  id_base = base_id(hostname = hostname),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)
}
\arguments{
\item{alias}{Alias of the data source. A character scalar.}

\item{type}{Type of the data source to add. One of \code{"mssql"}, \code{"mysql"}, \code{"pg"} or \code{"sqlite3"}.}

\item{config}{Type-specific configuration for the data source to add. A list.}

\item{inflection_column}{Type of inflection to apply for column names in the data source to add. One of
\code{"none"} or \code{"camelize"}.}

\item{inflection_table}{Type of inflection to apply for table names in the data source to add. One of
\code{"none"} or \code{"camelize"}.}

\item{enabled}{Whether the added data source is enabled or disabled.}

\item{id_base}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
\code{alias}, invisibly.
}
\description{
Adds a data source to the specified base on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/Source/operation/source-create}{\code{POST /api/v2/meta/bases/{id_base}/sources}} API endpoint.
}
\examples{
\dontrun{
nocodb::create_data_src(alias = "REPLACE-ME",
                        type = "pg",
                        config = list(client = "pg",
                                      connection = list(sslmode = "verify-full",
                                                        user = "REPLACE-ME",
                                                        password = "REPLACE-ME",
                                                        database = REPLACE-ME,
                                                        host = "REPLACE-ME",
                                                        port = 5432,
                                                        ssl = list(ca = "",
                                                                   cert = "",
                                                                   key = ""))),
                        inflection_column = "none",
                        inflection_table = "none")}
}
\seealso{
Other functions to manage NocoDB base data sources:
\code{\link{data_src}()},
\code{\link{data_src_id}()},
\code{\link{data_srcs}()},
\code{\link{delete_data_src}()}
}
\concept{data_src}
