% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{create_data_src_tbl}
\alias{create_data_src_tbl}
\title{Create table in NocoDB data source}
\usage{
create_data_src_tbl(
  id_data_src,
  name,
  cols,
  title = NULL,
  meta = NULL,
  order = NULL,
  id_base = base_id(origin = origin, email = email, password = password, api_token =
    api_token),
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)
}
\arguments{
\item{id_data_src}{\code{\link[base:character]{character(1)}}\cr
NocoDB data source identifier as returned by \code{\link[=data_src_id]{data_src_id()}}.}

\item{name}{\code{\link[base:character]{character(1)}}\cr
Table name.}

\item{cols}{\code{\link[base:list]{list()}}\cr
Table columns specification as a list or something coercible to like a data frame.}

\item{title}{\code{\link[base:character]{character(1)}}\cr
NocoDB-specific table title.}

\item{meta}{\code{\link[base:list]{list()}}\cr
NocoDB-specific table metadata as a list or something coercible to like a data frame.}

\item{order}{\code{\link[base:integer]{integer(1)}}\cr
A number to assign as the table's NocoDB-specific order "weight".}

\item{id_base}{\code{\link[base:character]{character(1)}}\cr
NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with metadata about the newly created NocoDB table, invisibly.
}
\description{
Adds a new table in the specified data source of the specified base on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/Sources/operation/table-create}{\code{POST /api/v2/meta/bases/{id_base}/{id_data_src}/tables}} API endpoint.
}
\examples{
\dontrun{
# create new table with a single PK column
nocodb::create_data_src_tbl(id_data_src = "REPLACE-ME",
                            name = "my_new_table",
                            title = "MyNewTable",
                            cols = list(list(column_name = "id",
                                             title = "id",
                                             dt = "text",
                                             dtx = "text",
                                             nrqd = FALSE,
                                             rqd = TRUE,
                                             ck = FALSE,
                                             pk = TRUE,
                                             un = FALSE,
                                             ai = FALSE,
                                             uidt = "LongText")))}
}
\seealso{
Other functions to manage NocoDB base data sources:
\code{\link{create_data_src}()},
\code{\link{data_src}()},
\code{\link{data_src_diff}()},
\code{\link{data_src_id}()},
\code{\link{data_src_tbls}()},
\code{\link{data_srcs}()},
\code{\link{delete_data_src}()},
\code{\link{has_data_src_diff}()},
\code{\link{sync_data_src}()},
\code{\link{test_data_src}()},
\code{\link{update_data_src}()}

Other functions to manage NocoDB tables:
\code{\link{create_tbl}()},
\code{\link{data_src_tbls}()},
\code{\link{delete_tbl}()},
\code{\link{reorder_tbl}()},
\code{\link{set_tbl_metadata}()},
\code{\link{tbl}},
\code{\link{tbl_id}()},
\code{\link{tbls}()},
\code{\link{update_tbl}()}
}
\concept{data_src}
\concept{tbls}
