% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{create_integration}
\alias{create_integration}
\title{Create NocoDB integration}
\usage{
create_integration(
  connection,
  title,
  type = "database",
  sub_type = c("mysql2", "pg", "sqlite3"),
  search_paths = "public",
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)
}
\arguments{
\item{connection}{\code{\link[base:list]{list()}}\cr
Subtype-specific connection details for database integrations.}

\item{title}{\code{pkgsnip::type("chr", 1L)}
NocoDB integration title.}

\item{type}{\code{pkgsnip::type("chr", 1L)}
NocoDB integration type. One of \code{"database"}, or \code{NULL} for any type.}

\item{sub_type}{\code{pkgsnip::type("chr", 1L)}
NocoDB integration subtype. One of \code{"mysql2"}, \code{"pg"} or \code{"sqlite3"}, or \code{NULL} for any subtype.}

\item{search_paths}{\code{\link[base:character]{character()}}\cr
Database search paths (schemas) to expose for the \code{"pg"} integration.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with metadata about the newly created NocoDB integration, invisibly.
}
\description{
Adds an integration on a NocoDB server via its \code{POST /api/v2/meta/integrations} API endpoint.
}
\details{
Note that for the database sub-type \code{sqlite3}, NocoDB silently creates a new SQLite database upon first use if none exists under the specified \code{filename} in
the \code{connection} list.
}
\examples{
\dontrun{
# create connection to PostgreSQL DB
nocodb::create_integration(connection = list(host = "REPLACE-ME",
                                             port = 5432,
                                             ssl = TRUE,
                                             sslmode = "verify-full",
                                             user = "REPLACE-ME",
                                             password = "REPLACE-ME",
                                             database = REPLACE-ME),
                           title = "Test PGSQL",
                           type = "database",
                           sub_type = "pg")

# create connection to local SQLite DB
nocodb::create_integration(connection = list(client = "sqlite3",
                                             database = "",
                                             connection = list(filename = "/test.db"),
                                             useNullAsDefault = TRUE),
                           title = "Test SQLite",
                           type = "database",
                           sub_type = "sqlite3")

}
}
\seealso{
Other functions to manage NocoDB integrations:
\code{\link{delete_integration}()},
\code{\link{integration}()},
\code{\link{integration_id}()},
\code{\link{integrations}()},
\code{\link{update_integration}()}
}
\concept{integrations}
