% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{create_tbl}
\alias{create_tbl}
\title{Create NocoDB table}
\usage{
create_tbl(
  name,
  cols,
  title = NULL,
  meta = NULL,
  order = NULL,
  id_base = base_id(hostname = hostname, email = email, password = password, api_token =
    api_token),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)
}
\arguments{
\item{name}{Table name. A character scalar.}

\item{cols}{Table columns specification as a list or something coercible to like a data frame.}

\item{title}{NocoDB-specific table title. A character scalar.}

\item{meta}{NocoDB-specific table metadata. A list.}

\item{order}{A number to assign as the table's NocoDB-specific order "weight".}

\item{id_base}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with metadata about the newly created NocoDB table, invisibly.
}
\description{
Adds a new table to the specified base's \emph{default} data source on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/DB-Table/operation/db-table-create}{\code{POST /api/v2/meta/bases/{id_base}/tables}} API endpoint.

To add a new table to a \emph{specific} (external) data source, use \code{\link[=create_data_src_tbl]{create_data_src_tbl()}} instead.
}
\details{
Besides the specified \code{cols}, NocoDB automatically creates the additional columns
\href{https://docs.nocodb.com/fields/field-types/date-time-based/created-time/}{\code{created_at}},
\href{https://docs.nocodb.com/fields/field-types/date-time-based/last-modified-time/}{\code{updated_at}},
\href{https://docs.nocodb.com/fields/field-types/user-based/created-by}{\code{created_by}} and
\href{https://docs.nocodb.com/fields/field-types/user-based/last-modified-by}{\code{updated_by}} as system fields.
}
\examples{
\dontrun{
# create new table with a single PK column
nocodb::create_tbl(name = "my_new_table",
                   title = "MyNewTable",
                   cols = list(list(column_name = "id",
                                    title = "id",
                                    dt = "text",
                                    dtx = "text",
                                    nrqd = FALSE,
                                    rqd = TRUE,
                                    ck = FALSE,
                                    pk = TRUE,
                                    un = FALSE,
                                    ai = FALSE,
                                    uidt = "LongText")))}
}
\seealso{
Other functions to manage NocoDB tables:
\code{\link{create_data_src_tbl}()},
\code{\link{data_src_tbls}()},
\code{\link{delete_tbl}()},
\code{\link{reorder_tbl}()},
\code{\link{set_tbl_metadata}()},
\code{\link{tbl}},
\code{\link{tbl_id}()},
\code{\link{tbls}()},
\code{\link{update_tbl}()}
}
\concept{tbls}
