% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{create_user}
\alias{create_user}
\title{Create NocoDB user}
\usage{
create_user(
  email,
  role = c("no-access", "commenter", "editor", "guest", "owner", "viewer", "creator"),
  base_id = base_id(),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  auth_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg),
  quiet = FALSE
)
}
\arguments{
\item{email}{E-mail address of the new user. A character scalar.}

\item{role}{Base role to assign to the new user. One of \code{"no-access"}, \code{"commenter"}, \code{"editor"}, \code{"guest"}, \code{"owner"}, \code{"viewer"} or \code{"creator"}.}

\item{base_id}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{auth_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API authentication token}. A character scalar.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{email}, invisibly.
}
\description{
Adds a new user account to the specified base on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/Auth/operation/auth-base-user-add}{\code{POST /api/v2/meta/bases/{base_id}/users}} API endpoint.
}
\seealso{
Other functions to manage NocoDB users:
\code{\link{users}()}
}
\concept{users}
