% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{delete_api_token}
\alias{delete_api_token}
\title{Delete NocoDB API token}
\usage{
delete_api_token(
  api_token,
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email", require = TRUE),
  password = pal::pkg_config_val("password", require = TRUE)
)
}
\arguments{
\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB API token to delete.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}
}
\value{
\code{api_token}, invisibly.
}
\description{
Deletes the specified API token on a NocoDB server via its
\href{https://docs.nocodb.com/0.109.7/developer-resources/rest-apis/#meta-apis}{\code{DELETE /api/v1/tokens/{api_token}}} API endpoint.
}
\details{
This API endpoint does not support authentication via \href{https://docs.nocodb.com/account-settings/api-tokens/}{API tokens}.
}
\seealso{
Other functions to manage NocoDB API tokens:
\code{\link{api_tokens}()},
\code{\link{create_api_token}()}

Other functions to manage NocoDB authentication:
\code{\link{api_tokens}()},
\code{\link{assert_super_admin}()},
\code{\link{create_api_token}()},
\code{\link{is_signed_in}()},
\code{\link{is_super_admin}()},
\code{\link{refresh_sign_in}()},
\code{\link{req_auth}()},
\code{\link{sign_in}()},
\code{\link{sign_out}()},
\code{\link{sign_up_user}()}
}
\concept{api_tokens}
\concept{auth}
