% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{integration}
\alias{integration}
\title{Get NocoDB integration}
\usage{
integration(
  id_integration,
  incl_config = TRUE,
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)
}
\arguments{
\item{id_integration}{\code{\link[base:character]{character(1)}}\cr
NocoDB integration identifier as returned by \code{\link[=integration_id]{integration_id()}}.}

\item{incl_config}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to include the integration configuration in the result (as a separate \code{config} column).}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with metadata about the specified integration on a NocoDB server from its \code{GET /api/v2/meta/integrations/{id_integration}}
API endpoint.
}
\seealso{
Other functions to manage NocoDB integrations:
\code{\link{create_integration}()},
\code{\link{delete_integration}()},
\code{\link{integration_id}()},
\code{\link{integrations}()},
\code{\link{update_integration}()}
}
\concept{integrations}
