% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{integration_id}
\alias{integration_id}
\title{Get NocoDB integration ID}
\usage{
integration_id(
  title = NULL,
  type = "database",
  sub_type = c("mysql2", "pg", "sqlite3"),
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)
}
\arguments{
\item{title}{\code{pkgsnip::type("chr", 1L)}
NocoDB integration title. \code{NULL} means any title.}

\item{type}{\code{pkgsnip::type("chr", 1L)}
NocoDB integration type. One of \code{"database"}, or \code{NULL} for any type.}

\item{sub_type}{\code{pkgsnip::type("chr", 1L)}
NocoDB integration subtype. One of \code{"mysql2"}, \code{"pg"} or \code{"sqlite3"}, or \code{NULL} for any subtype.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
A character scalar.
}
\description{
Returns the identifier of the integration with the specified \code{type}, \code{sub_type} and/or \code{title} on a NocoDB server.
}
\details{
Note that NocoDB allows to create multiple integrations with the same \code{title}, \code{type} and \code{sub_type}, so \code{integration_id()} cannot discern such cases and
just returns the identifier of the \emph{first} integration listed in the API response (with a suitable warning).
}
\seealso{
Other functions to manage NocoDB integrations:
\code{\link{create_integration}()},
\code{\link{delete_integration}()},
\code{\link{integration}()},
\code{\link{integrations}()},
\code{\link{update_integration}()}
}
\concept{integrations}
