% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{integrations}
\alias{integrations}
\title{List NocoDB integrations}
\usage{
integrations(
  type = NULL,
  id_base = NULL,
  incl_config = TRUE,
  decode_config = TRUE,
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token")
)
}
\arguments{
\item{type}{\code{\link[base:character]{character(1)}}\cr
Optional integration type to limit the results to. One of \code{"database"}.}

\item{id_base}{\code{\link[base:character]{character(1)}}\cr
Optional NocoDB base identifier (as returned by \code{\link[=base_id]{base_id()}}) to limit the results to.}

\item{incl_config}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to include the integration configuration in the result.}

\item{decode_config}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to decode the integration configuration in the result. If \code{FALSE}, the whole configuration is returned as a single Base64-encoded and
encrypted \code{config} column, otherwise it is returned as multiple clear-text \verb{config.*} columns. Only relevant if \code{incl_config = TRUE}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with metadata about all configured integrations on a NocoDB server from its \code{GET /api/v2/meta/integrations} API endpoint.
}
\seealso{
Other functions to manage NocoDB integrations:
\code{\link{create_integration}()},
\code{\link{delete_integration}()},
\code{\link{integration}()},
\code{\link{integration_id}()},
\code{\link{update_integration}()}
}
\concept{integrations}
