% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{is_access_token_expired}
\alias{is_access_token_expired}
\title{Test if access token is expired}
\usage{
is_access_token_expired(x)
}
\arguments{
\item{x}{Access token to be tested for expiration.}
}
\value{
A logical scalar.
}
\description{
Tests whether an \link[=access_token]{access token} is still valid or not.
}
\seealso{
Other access_token: 
\code{\link{access_token}()},
\code{\link{decode_access_token}()},
\code{\link{store_access_token}()}
}
\concept{access_token}
\keyword{internal}
