% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{is_plugin_active}
\alias{is_plugin_active}
\title{Test if plugin is active}
\usage{
is_plugin_active(
  title,
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token")
)
}
\arguments{
\item{title}{\code{\link[base:character]{character(1)}}\cr
NocoDB plugin title.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
A logical scalar.
}
\description{
Tests whether the plugin with the specified \code{title} is active on a NocoDB server via its
\href{https://docs.nocodb.com/0.109.7/developer-resources/rest-apis/#meta-apis}{\code{GET /api/v1/db/meta/plugins/{title}/status}} API endpoint.
}
\details{
Note that the API endpoint doesn't check the plugin \code{title} for validity and simply returns \code{FALSE} for a \code{title} which does not actually exist.
}
\seealso{
Other functions to manage NocoDB plugins:
\code{\link{plugin}()},
\code{\link{plugin_category}()},
\code{\link{plugin_id}()},
\code{\link{plugins}()},
\code{\link{test_plugin}()},
\code{\link{update_plugin}()}
}
\concept{plugins}
