% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb-package.R
\docType{package}
\name{nocodb-package}
\alias{nocodb}
\alias{nocodb-package}
\title{nocodb: Use NocoDB's RESTful APIs}
\description{
nocodb allows to work with NocoDB's RESTful APIs from R. NocoDB (https://nocodb.com) is a web application that provides an intuitive spreadsheet-like front-end for relational databases.
}
\section{Package configuration}{


Some of nocodb's functionality is controlled via package-specific global configuration which can either be set via \href{https://rdrr.io/r/base/options.html}{R options} or \href{https://en.wikipedia.org/wiki/Environment_variable}{environment variables} (the former take precedence). This configuration includes:\tabular{llll}{
   \strong{Description} \tab \strong{R option} \tab \strong{Environment variable} \tab \strong{Default value} \cr
   \href{https://en.wikipedia.org/wiki/Hostname}{Hostname} of the NocoDB server to work with. \tab \code{nocodb.hostname} \tab \code{R_NOCODB_HOSTNAME} \tab  \cr
   Title (name) of the NocoDB \href{https://docs.nocodb.com/category/bases}{base} to work with. \tab \code{nocodb.base_title} \tab \code{R_NOCODB_BASE_TITLE} \tab  \cr
   E-mail address of the NocoDB user to authenticate with. \tab \code{nocodb.email} \tab \code{R_NOCODB_EMAIL} \tab  \cr
   Password of the NocoDB user to authenticate with. \tab \code{nocodb.password} \tab \code{R_NOCODB_PASSWORD} \tab  \cr
   \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token} belonging to the NocoDB user to work with. An alternative to \code{email} and \code{password}. Note that not all of NocoDB's API endpoints support authentication via API tokens. \tab \code{nocodb.api_token} \tab \code{R_NOCODB_API_TOKEN} \tab  \cr
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://nocodb.rpkg.dev}
  \item \url{https://gitlab.com/rpkg.dev/nocodb}
  \item Report bugs at \url{https://gitlab.com/rpkg.dev/nocodb/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
