% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{plugin_category}
\alias{plugin_category}
\title{Get NocoDB plugin category}
\usage{
plugin_category(
  id_plugin,
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)
}
\arguments{
\item{id_plugin}{\code{\link[base:character]{character(1)}}\cr
NocoDB plugin identifier as returned by \code{\link[=plugin_id]{plugin_id()}}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
Plugin category as a character scalar.
}
\description{
Returns the category of the specified plugin on a NocoDB server.
}
\seealso{
Other functions to manage NocoDB plugins:
\code{\link{is_plugin_active}()},
\code{\link{plugin}()},
\code{\link{plugin_id}()},
\code{\link{plugins}()},
\code{\link{test_plugin}()},
\code{\link{update_plugin}()}
}
\concept{plugins}
