% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{refresh_access_token}
\alias{refresh_access_token}
\title{Refresh NocoDB access token}
\usage{
refresh_access_token(
  email = pal::pkg_config_val(key = "email", pkg = this_pkg),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg)
)
}
\arguments{
\item{email}{E-mail address of the NocoDB user whose access token is to be refreshed. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}
}
\value{
The freshly generated access token as a character scalar, invisibly.
}
\description{
Requests a fresh NocoDB access token via the \href{https://data-apis-v1.nocodb.com/#tag/Auth/operation/auth-token-refresh}{\code{POST /api/v1/auth/token/refresh}} API
endpoint. Requires a valid refresh token cached to disk by \code{\link[=sign_in]{sign_in(cache_refresh_token = TRUE)}}.
}
\details{
Besides generating a fresh access token, also a new refresh token is cached to disk.
}
\seealso{
Other common functions:
\code{\link{api}()},
\code{\link{req_auth}()},
\code{\link{req_basic}()},
\code{\link{sign_in}()},
\code{\link{sign_out}()}

Other functions to manage NocoDB authentication:
\code{\link{req_auth}()},
\code{\link{sign_in}()},
\code{\link{sign_out}()}
}
\concept{auth}
\concept{common}
\keyword{internal}
