% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{reorder_tbl}
\alias{reorder_tbl}
\title{Re-order NocoDB table}
\usage{
reorder_tbl(
  id_tbl,
  order = 1L,
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)
}
\arguments{
\item{id_tbl}{\code{\link[base:character]{character(1)}}\cr
NocoDB table identifier as returned by \code{\link[=tbl_id]{tbl_id()}}.}

\item{order}{\code{\link[base:integer]{integer(1)}}\cr
A number to assign as the table's NocoDB-specific order "weight".}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
\code{id_tbl}, invisibly.
}
\description{
Sets the numeric order of the specified table on a NocoDB server via its
\code{POST /api/v2/meta/tables/{id_tbl}/reorder} API endpoint.
}
\details{
Lower numbers place the table higher up in the UI and vice versa. The current order of all the tables in a base can be determined via \code{\link[=tbls]{tbls()}}.
}
\seealso{
Other functions to manage NocoDB tables:
\code{\link{create_data_src_tbl}()},
\code{\link{create_tbl}()},
\code{\link{data_src_tbls}()},
\code{\link{delete_tbl}()},
\code{\link{set_tbl_metadata}()},
\code{\link{tbl}},
\code{\link{tbl_id}()},
\code{\link{tbls}()},
\code{\link{update_tbl}()}
}
\concept{tbls}
