% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{req_basic}
\alias{req_basic}
\title{Create basic NocoDB API request}
\usage{
req_basic(
  path,
  method = c("GET", "CONNECT", "DELETE", "HEAD", "OPTIONS", "PATCH", "POST", "PUT",
    "TRACE"),
  origin = pal::pkg_config_val("origin"),
  max_tries = 3L
)
}
\arguments{
\item{path}{\code{\link[base:character]{character(1)}}\cr
NocoDB API endpoint path.}

\item{method}{\code{\link[base:character]{character(1)}}\cr
\href{https://developer.mozilla.org/en-US/docs/Web/HTTP/Methods}{HTTP request method}.
One of \code{"GET"}, \code{"CONNECT"}, \code{"DELETE"}, \code{"HEAD"}, \code{"OPTIONS"}, \code{"PATCH"}, \code{"POST"}, \code{"PUT"} or \code{"TRACE"}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{max_tries}{Maximum number of attempts to retry the request in case of an HTTP error. An integerish scalar. The retries are performed using exponential
backoff and jitter, see \code{\link[httr2:req_retry]{httr2::req_retry()}} for details.}
}
\value{
A modified HTTP \link[httr2]{request}.
}
\description{
Assembles the HTTP request structure that is common to \emph{all} NocoDB API requests performed by this package.
}
\seealso{
Other common functions:
\code{\link{api}()},
\code{\link{req_auth}()}
}
\concept{common}
\keyword{internal}
