% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{set_tbl_metadata}
\alias{set_tbl_metadata}
\title{Set metadata for NocoDB tables}
\usage{
set_tbl_metadata(
  data,
  base_id = base_id(),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  auth_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg),
  quiet = FALSE
)
}
\arguments{
\item{data}{Dataframe with the columns \code{name}, \code{meta.icon} that defines the table-name-and-metadata-value mapping to be applied. Additional columns are
ignored.}

\item{base_id}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{auth_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API authentication token}. A character scalar.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Sets the provided table metadata on a NocoDB server. Currently, this includes:
\itemize{
\item Setting the order of the tables in the base using \code{\link[=reorder_tbl]{reorder_tbl()}} according to the row order of \code{data}.
\item Setting the \href{https://docs.nocodb.com/tables/actions-on-table/#change-table-icon}{table icons} (emojis) using \code{\link[=update_tbl]{update_tbl()}} according to \code{data$meta.icon}.
}
}
\seealso{
Other functions to manage NocoDB tables:
\code{\link{reorder_tbl}()},
\code{\link{tbl}},
\code{\link{tbl_id}()},
\code{\link{tbls}()},
\code{\link{update_tbl}()}
}
\concept{tbls}
