% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{sign_up_user}
\alias{sign_up_user}
\title{Sign up NocoDB user}
\usage{
sign_up_user(
  email_new,
  password_new,
  invite_token = NULL,
  subscribe_to_newsletter = FALSE,
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg, required = FALSE),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg, required = FALSE),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg, required = FALSE)
)
}
\arguments{
\item{email_new}{E-mail address of the user to sign up. A character scalar.}

\item{password_new}{Password of the user to sign up. A character scalar.}

\item{invite_token}{Invite token of the user to sign up. A character scalar.}

\item{subscribe_to_newsletter}{Whether or not to subscribe the signed up user to the NocoDB newsletter.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
The generated access token for the signed up user as a character scalar, invisibly.
}
\description{
Signs up a user on a NocoDB server via its \href{https://data-apis-v1.nocodb.com/#tag/Auth/operation/auth-signup}{\code{POST /api/v1/auth/user/signup}} API endpoint.
If sign-up is \href{https://docs.nocodb.com/account-settings/oss-specific-details/#enable--disable-signup}{restricted to invitees only}, the user must have been
\link[=invite_user]{invited} before and an \code{invite_token} must be provided. Otherwise, the user is created straightaway (without sending an e-mail).
}
\section{API errors}{
\itemize{
\item If you \emph{did} provide an \code{invite_token} and encounter a \strong{\verb{Not allowed to signup}} error, it means the provided \code{email_new} does not match the
\code{invite_token}.
\item If you \emph{did not} provide an \code{invite_token} and encounter a \strong{\verb{Not allowed to signup}} error, it means sign-up of new users is \href{https://docs.nocodb.com/account-settings/oss-specific-details/#enable--disable-signup}{restricted to invitees only}.
\item If you encounter an \strong{\verb{Invalid invite url}} error, it means the provided \code{invite_token} is either invalid or has already been used to sign up.
}
}

\seealso{
Other functions to manage NocoDB users:
\code{\link{base_users}()},
\code{\link{invite_user}()},
\code{\link{update_base_user}()},
\code{\link{update_user}()},
\code{\link{user}()},
\code{\link{user_id}()}

Other functions to manage NocoDB authentication:
\code{\link{is_signed_in}()},
\code{\link{refresh_sign_in}()},
\code{\link{req_auth}()},
\code{\link{sign_in}()},
\code{\link{sign_out}()}
}
\concept{auth}
\concept{users}
