% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{store_access_token}
\alias{store_access_token}
\title{Store NocoDB access token}
\usage{
store_access_token(
  x,
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg)
)
}
\arguments{
\item{x}{Access token to be stored.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}
}
\value{
\code{x}, invisibly.
}
\description{
Stores a NocoDB user's access token in the \code{stateful} package environment.
}
\seealso{
Other access_token: 
\code{\link{access_token}()},
\code{\link{decode_access_token}()},
\code{\link{is_access_token_expired}()}
}
\concept{access_token}
\keyword{internal}
