% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{sync_data_src}
\alias{sync_data_src}
\title{Synchronize NocoDB data source schema}
\usage{
sync_data_src(
  id_data_src,
  id_base = base_id(origin = origin, email = email, password = password),
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password")
)
}
\arguments{
\item{id_data_src}{\code{\link[base:character]{character(1)}}\cr
NocoDB data source identifier as returned by \code{\link[=data_src_id]{data_src_id()}}.}

\item{id_base}{\code{\link[base:character]{character(1)}}\cr
NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}
}
\value{
\code{id_data_src}, invisibly.
}
\description{
Synchronizes the specified data source's schema between its external and its NocoDB-internal state via the
\code{POST /api/v2/meta/bases/{id_base}/meta-diff/{id_data_src}} API endpoint.
}
\details{
This API endpoint does not support authentication via \href{https://docs.nocodb.com/account-settings/api-tokens/}{API tokens}.
}
\seealso{
Other functions to manage NocoDB base data sources:
\code{\link{create_data_src}()},
\code{\link{create_data_src_tbl}()},
\code{\link{data_src}()},
\code{\link{data_src_diff}()},
\code{\link{data_src_id}()},
\code{\link{data_src_tbls}()},
\code{\link{data_srcs}()},
\code{\link{delete_data_src}()},
\code{\link{has_data_src_diff}()},
\code{\link{test_data_src}()},
\code{\link{update_data_src}()}
}
\concept{data_src}
