% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{tbl_col}
\alias{tbl_col}
\title{Get NocoDB table column metadata}
\usage{
tbl_col(
  col_id,
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  auth_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)
}
\arguments{
\item{col_id}{NocoDB column identifier as returned by \code{\link[=tbl_col_id]{tbl_col_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{auth_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API authentication token}. A character scalar.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with metadata about the specified column on a NocoDB server from its
\href{https://meta-apis-v2.nocodb.com/#tag/DB-Table-Column/operation/db-table-column-get}{\code{GET /api/v2/meta/columns/{col_id}}} API endpoint.
}
\seealso{
Other functions to manage NocoDB table columns:
\code{\link{set_display_val}()},
\code{\link{set_display_vals}()},
\code{\link{tbl_col_id}()},
\code{\link{tbl_cols}()},
\code{\link{update_tbl_col}()}
}
\concept{cols}
