% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{tbl_cols}
\alias{tbl_cols}
\title{List NocoDB table columns}
\usage{
tbl_cols(
  id_tbl,
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)
}
\arguments{
\item{id_tbl}{\code{\link[base:character]{character(1)}}\cr
NocoDB table identifier as returned by \code{\link[=tbl_id]{tbl_id()}}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with metadata about the columns of the specified table on a NocoDB server from its
\href{https://meta-apis-v2.nocodb.com/#tag/DB-Table/operation/db-table-read}{\code{GET /api/v2/meta/tables/{id_tbl}}} API endpoint.
}
\seealso{
Other functions to manage NocoDB table columns:
\code{\link{create_tbl_col}()},
\code{\link{delete_tbl_col}()},
\code{\link{set_display_val}()},
\code{\link{set_display_vals}()},
\code{\link{tbl_col}()},
\code{\link{tbl_col_id}()},
\code{\link{update_tbl_col}()}
}
\concept{cols}
