% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{tbl_view_col_id}
\alias{tbl_view_col_id}
\title{Get NocoDB table view column ID}
\usage{
tbl_view_col_id(
  id_tbl,
  name = NULL,
  title = NULL,
  view_type = "default",
  view_name = NULL,
  view_title = NULL,
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)
}
\arguments{
\item{id_tbl}{\code{\link[base:character]{character(1)}}\cr
NocoDB table identifier as returned by \code{\link[=tbl_id]{tbl_id()}}.}

\item{name}{\code{\link[base:character]{character(1)}}\cr
NocoDB column name. \code{NULL} means any name.}

\item{title}{\code{\link[base:character]{character(1)}}\cr
NocoDB column title. \code{NULL} means any title.}

\item{view_type}{\code{\link[base:character]{character(1)}}\cr
NocoDB view type. One of \code{"default"}, \code{"form"}, \code{"gallery"}, \code{"grid"}, \code{"kanban"}, \code{"map"}, \code{"calendar"}, or \code{NULL} for any type.}

\item{view_name}{\code{\link[base:character]{character(1)}}\cr
NocoDB view name. \code{NULL} means any name.}

\item{view_title}{\code{\link[base:character]{character(1)}}\cr
NocoDB view title. \code{NULL} means any title.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
A character scalar.
}
\description{
Returns the identifier of the view column with the specified \code{name} or \code{title} in the table with the specified \code{id_tbl} on a NocoDB server.
}
\seealso{
Other functions to manage NocoDB table views:
\code{\link{base_ui_acl}()},
\code{\link{tbl_view_cols}()},
\code{\link{tbl_view_id}()},
\code{\link{tbl_views}()},
\code{\link{update_base_ui_acl}()},
\code{\link{update_tbl_view_col}()}
}
\concept{views}
