% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{tbl_view_cols}
\alias{tbl_view_cols}
\title{List NocoDB table view columns}
\usage{
tbl_view_cols(
  id_view,
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)
}
\arguments{
\item{id_view}{\code{\link[base:character]{character(1)}}\cr
NocoDB table view identifier as returned by \code{\link[=tbl_view_id]{tbl_view_id()}}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with metadata about the columns of the specified table view on a NocoDB server from its
\href{https://meta-apis-v2.nocodb.com/#tag/Views/operation/db-view-column-list}{\code{GET /api/v2/meta/views/{id_view}/columns}} API endpoint.
}
\seealso{
Other functions to manage NocoDB table views:
\code{\link{base_ui_acl}()},
\code{\link{tbl_view_col_id}()},
\code{\link{tbl_view_id}()},
\code{\link{tbl_views}()},
\code{\link{update_base_ui_acl}()},
\code{\link{update_tbl_view_col}()}
}
\concept{views}
