% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{tbl_view_id}
\alias{tbl_view_id}
\title{Get NocoDB table view ID}
\usage{
tbl_view_id(
  id_tbl,
  type = "default",
  name = NULL,
  title = NULL,
  hostname = pal::pkg_config_val("hostname"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token")
)
}
\arguments{
\item{id_tbl}{NocoDB table identifier as returned by \code{\link[=tbl_id]{tbl_id()}}. A character scalar.}

\item{type}{NocoDB view type. One of \code{"default"}, \code{"form"}, \code{"gallery"}, \code{"grid"}, \code{"kanban"}, \code{"map"}, \code{"calendar"}, or \code{NULL} for any type.}

\item{name}{NocoDB view name. A character scalar, or \code{NULL} for any name.}

\item{title}{NocoDB view title. A character scalar, or \code{NULL} for any title.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
A character scalar.
}
\description{
Returns the identifier of the view with the specified \code{type}, \code{name} and/or \code{title} in the table with the specified \code{id_tbl} on a NocoDB server.
}
\seealso{
Other functions to manage NocoDB table views:
\code{\link{base_ui_acl}()},
\code{\link{tbl_views}()},
\code{\link{update_base_ui_acl}()}
}
\concept{views}
