% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{tbl_views}
\alias{tbl_views}
\title{List NocoDB table views}
\usage{
tbl_views(
  id_tbl,
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token")
)
}
\arguments{
\item{id_tbl}{\code{\link[base:character]{character(1)}}\cr
NocoDB table identifier as returned by \code{\link[=tbl_id]{tbl_id()}}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with metadata about views of specified NocoDB table.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with metadata about the views of the specified table on a NocoDB server from its
\href{https://meta-apis-v2.nocodb.com/#tag/DB-View/operation/db-view-list}{\code{GET /api/v2/meta/tables/{id_tbl}/views}} API endpoint.
}
\seealso{
Other functions to manage NocoDB table views:
\code{\link{base_ui_acl}()},
\code{\link{tbl_view_id}()},
\code{\link{update_base_ui_acl}()}
}
\concept{views}
