% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{test_plugin}
\alias{test_plugin}
\title{Test NocoDB plugin configuration}
\usage{
test_plugin(
  title,
  config,
  category = plugin_category(id_plugin = plugin_id(title = title, origin = origin, email
    = email, password = password, api_token = api_token), origin = origin, email = email,
    password = password, api_token = api_token),
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token")
)
}
\arguments{
\item{title}{\code{\link[base:character]{character(1)}}\cr
NocoDB plugin title.}

\item{config}{\code{\link[base:list]{list()}}\cr
Plugin configuration.}

\item{category}{\code{\link[base:character]{character(1)}}\cr
NocoDB plugin category as returned by \code{\link[=plugin_category]{plugin_category()}}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
\code{TRUE} if the test was successful.
}
\description{
Tests the configuration of the plugin with the specified title and category on a NocoDB server via its
\href{https://docs.nocodb.com/0.109.7/developer-resources/rest-apis/#meta-apis}{\code{POST /api/v1/db/meta/plugins/test}} API endpoint.
}
\examples{
\dontrun{
nocodb::test_plugin(title = "Backblaze B2",
                    config = list(bucket = "REPLACE-ME",
                                  region = "REPLACE-ME",
                                  access_key = "REPLACE-ME",
                                  access_secret = "REPLACE-ME"))

nocodb::test_plugin(title = "S3",
                    config = list(endpoint = "https://s3.REPLACE-ME.backblazeb2.com",
                                  region = "REPLACE-ME",
                                  bucket = "REPLACE-ME",
                                  access_key = "REPLACE-ME",
                                  access_secret = "REPLACE-ME",
                                  acl = "public-read"))}
}
\seealso{
Other functions to manage NocoDB plugins:
\code{\link{is_plugin_active}()},
\code{\link{plugin}()},
\code{\link{plugin_category}()},
\code{\link{plugin_id}()},
\code{\link{plugins}()},
\code{\link{update_plugin}()}
}
\concept{plugins}
