% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{update_base_user}
\alias{update_base_user}
\title{Update NocoDB base user}
\usage{
update_base_user(
  user_email,
  role = c("no-access", "viewer", "commenter", "editor", "creator"),
  id_user = user_id(user_email = user_email, id_base = id_base, origin = origin, email =
    email, password = password, api_token = api_token),
  id_base = base_id(origin = origin, email = email, password = password, api_token =
    api_token),
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token"),
  api_version = c("v2", "v1"),
  quiet = FALSE
)
}
\arguments{
\item{user_email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the user.}

\item{role}{\code{\link[base:character]{character(1)}}\cr
\href{https://docs.nocodb.com/collaboration/oss-specific-details/#base-level-permissions}{Base role} to assign to the user. One of
\code{"no-access"}, \code{"viewer"}, \code{"commenter"}, \code{"editor"} or \code{"creator"}.}

\item{id_user}{\code{\link[base:character]{character(1)}}\cr
NocoDB user identifier as returned by \code{\link[=user_id]{user_id()}}.}

\item{id_base}{\code{\link[base:character]{character(1)}}\cr
NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}

\item{api_version}{\code{\link[base:character]{character(1)}}\cr
API endpoint version to use. Either
\itemize{
\item \code{"v2"} for \href{https://docs.nocodb.com/0.109.7/developer-resources/rest-apis/#meta-apis}{\code{PATCH /api/v1/db/meta/projects/{id_base}/users/{id_user}}}, or
\item \code{"v1"} for \href{https://docs.nocodb.com/0.109.7/developer-resources/rest-apis/#meta-apis}{\code{PATCH /api/v1/db/meta/projects/{id_base}/users/{id_user}}}.
}}

\item{quiet}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{id_user}, invisibly.
}
\description{
Updates the specified user in regard to the specified base on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/Auth/operation/auth-base-user-update}{\code{PATCH /api/v2/meta/bases/{id_base}/users/{id_user}}} or
\href{https://docs.nocodb.com/0.109.7/developer-resources/rest-apis/#meta-apis}{\code{PATCH /api/v1/db/meta/projects/{id_base}/users/{id_user}}} API endpoint,
depending on \code{api_version}.
}
\details{
Only the super admin user is allowed to use this API endpoint, i.e. the provided credentials must belong to them.

Note that for the sake of convenience, this function automatically falls back to \code{\link[=invite_base_user]{invite_base_user()}} if necessary (i.e. the specified user hasn't yet been
invited to the specified base).
}
\seealso{
Other functions to manage NocoDB users:
\code{\link{add_user}()},
\code{\link{base_users}()},
\code{\link{delete_base_user}()},
\code{\link{delete_user}()},
\code{\link{invite_base_user}()},
\code{\link{invite_user}()},
\code{\link{resend_base_user_invitation}()},
\code{\link{sign_up_user}()},
\code{\link{update_user}()},
\code{\link{user_id}()},
\code{\link{users}()},
\code{\link{validate_user_email}()},
\code{\link{whoami}()}

Other functions to manage NocoDB bases:
\code{\link{base}()},
\code{\link{base_id}()},
\code{\link{base_ui_acl}()},
\code{\link{base_users}()},
\code{\link{bases}()},
\code{\link{create_base}()},
\code{\link{delete_base}()},
\code{\link{delete_base_user}()},
\code{\link{invite_base_user}()},
\code{\link{resend_base_user_invitation}()},
\code{\link{update_base}()},
\code{\link{update_base_ui_acl}()}
}
\concept{bases}
\concept{users}
