% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{update_grid_view_col}
\alias{update_grid_view_col}
\title{Update NocoDB grid view column}
\usage{
update_grid_view_col(
  id_view_col,
  order = NULL,
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)
}
\arguments{
\item{id_view_col}{\code{\link[base:character]{character(1)}}\cr
NocoDB table view column identifier as returned by \code{\link[=tbl_view_col_id]{tbl_view_col_id()}}.}

\item{order}{\code{\link[base:integer]{integer(1)}}\cr
A number to assign as the grid view columns's UI order "weight".}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
\code{id_view_col}, invisibly.
}
\description{
Updates the metadata of the specified grid view column on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/Views/operation/db-view-grid-column-update}{\code{PATCH /api/v2/meta/grid-columns/{id_view_col}}} API endpoint.
}
\seealso{
Other functions to manage NocoDB table views:
\code{\link{base_ui_acl}()},
\code{\link{tbl_view_col_id}()},
\code{\link{tbl_view_cols}()},
\code{\link{tbl_view_id}()},
\code{\link{tbl_views}()},
\code{\link{update_base_ui_acl}()},
\code{\link{update_tbl_view_col}()}
}
\concept{views}
