% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{update_integration}
\alias{update_integration}
\title{Update NocoDB integration}
\usage{
update_integration(
  id_integration,
  connection = NULL,
  title = NULL,
  type = NULL,
  sub_type = NULL,
  search_paths = NULL,
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token")
)
}
\arguments{
\item{id_integration}{\code{\link[base:character]{character(1)}}\cr
NocoDB integration identifier as returned by \code{\link[=integration_id]{integration_id()}}.}

\item{connection}{\code{\link[base:list]{list()}}\cr
Subtype-specific connection details for database integrations.}

\item{title}{\code{pkgsnip::type("chr", 1L)}
NocoDB integration title.}

\item{type}{\code{pkgsnip::type("chr", 1L)}
NocoDB integration type. One of \code{"database"}, or \code{NULL} for any type.}

\item{sub_type}{\code{pkgsnip::type("chr", 1L)}
NocoDB integration subtype. One of \code{"mysql2"}, \code{"pg"} or \code{"sqlite3"}, or \code{NULL} for any subtype.}

\item{search_paths}{\code{\link[base:character]{character()}}\cr
Database search paths (schemas) to expose for the \code{"pg"} integration.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with metadata about the updated NocoDB integration, invisibly.
}
\description{
Updates an integration on a NocoDB server via its \code{PATCH /api/v2/meta/integrations} API endpoint.
}
\seealso{
Other functions to manage NocoDB integrations:
\code{\link{create_integration}()},
\code{\link{delete_integration}()},
\code{\link{integration}()},
\code{\link{integration_id}()},
\code{\link{integrations}()}
}
\concept{integrations}
