% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{update_plugin}
\alias{update_plugin}
\title{Update NocoDB plugin configuration}
\usage{
update_plugin(
  id_plugin,
  config = NULL,
  activate = NULL,
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)
}
\arguments{
\item{id_plugin}{\code{\link[base:character]{character(1)}}\cr
NocoDB plugin identifier as returned by \code{\link[=plugin_id]{plugin_id()}}.}

\item{config}{\code{\link[base:list]{list()}}\cr
Plugin configuration. If \code{NULL}, the plugin's configuration is left untouched.}

\item{activate}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to activate the plugin. If \code{NULL}, the plugin's activation status is left untouched.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with metadata about the updated NocoDB plugin, invisibly.
}
\description{
Updates the configuration of the specified plugin on a NocoDB server via its
\href{https://docs.nocodb.com/0.109.7/developer-resources/rest-apis/#meta-apis}{\code{PATCH /api/v1/db/meta/plugins/{id_plugin}}} API endpoint.
}
\examples{
\dontrun{
nocodb::update_plugin(id_plugin = nocodb::plugin_id(title = "Backblaze B2"),
                      config = list(region = "REPLACE-ME",
                                    bucket = "REPLACE-ME",
                                    access_key = "REPLACE-ME",
                                    access_secret = "REPLACE-ME"),
                      activate = TRUE)
                      
nocodb::update_plugin(id_plugin = nocodb::plugin_id(title = "S3"),
                      config = list(endpoint = "https://s3.REPLACE-ME.backblazeb2.com",
                                    region = "REPLACE-ME",
                                    bucket = "REPLACE-ME",
                                    access_key = "REPLACE-ME",
                                    access_secret = "REPLACE-ME",
                                    acl = "public-read"),
                      activate = TRUE)}
}
\seealso{
Other functions to manage NocoDB plugins:
\code{\link{is_plugin_active}()},
\code{\link{plugin}()},
\code{\link{plugin_category}()},
\code{\link{plugin_id}()},
\code{\link{plugins}()},
\code{\link{test_plugin}()}
}
\concept{plugins}
