% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{update_tbl_view_col}
\alias{update_tbl_view_col}
\title{Update NocoDB table view column}
\usage{
update_tbl_view_col(
  id_view_col,
  id_view,
  show = NULL,
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)
}
\arguments{
\item{id_view_col}{\code{\link[base:character]{character(1)}}\cr
NocoDB table view column identifier as returned by \code{\link[=tbl_view_col_id]{tbl_view_col_id()}}.}

\item{id_view}{\code{\link[base:character]{character(1)}}\cr
NocoDB table view identifier as returned by \code{\link[=tbl_view_id]{tbl_view_id()}}.}

\item{show}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to show the column in the specified view. Omitted if \code{NULL}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
\code{id_view_col}, invisibly.
}
\description{
Updates the metadata of the specified table view column on a NocoDB server via its
\code{PATCH /api/v2/meta/views/{id_view}/columns/{id_view_col}} API endpoint.
}
\seealso{
Other functions to manage NocoDB table views:
\code{\link{base_ui_acl}()},
\code{\link{tbl_view_col_id}()},
\code{\link{tbl_view_cols}()},
\code{\link{tbl_view_id}()},
\code{\link{tbl_views}()},
\code{\link{update_base_ui_acl}()},
\code{\link{update_grid_view_col}()}
}
\concept{views}
