% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{update_user}
\alias{update_user}
\title{Update NocoDB user}
\usage{
update_user(
  display_name = NULL,
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = NULL,
  quiet = TRUE
)
}
\arguments{
\item{display_name}{\code{\link[base:character]{character(1)}}\cr
Name to be displayed for the user in NocoDB. Omitted if \code{NULL}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}

\item{quiet}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to suppress printing status output from internal processing.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with metadata about the updated NocoDB user, invisibly.
}
\description{
Updates the metadata of the specified user on a NocoDB server via its \code{PATCH /api/v1/user/profile} API endpoint.
}
\details{
The user is determined based on \code{api_token} or \code{email} and \code{password} (the former takes precedence).

In order for other API endpoints to reflect the updated metadata, NocoDB's internal state must update, which is only triggered by this function iff the user
to be updated is the super admin and authentication is done via \code{email} and \code{password} instead of \code{api_token}.
}
\seealso{
Other functions to manage NocoDB users:
\code{\link{add_user}()},
\code{\link{base_users}()},
\code{\link{delete_base_user}()},
\code{\link{delete_user}()},
\code{\link{invite_base_user}()},
\code{\link{invite_user}()},
\code{\link{resend_base_user_invitation}()},
\code{\link{sign_up_user}()},
\code{\link{update_base_user}()},
\code{\link{user_id}()},
\code{\link{users}()},
\code{\link{validate_user_email}()},
\code{\link{whoami}()}
}
\concept{users}
