% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{upload_attachments}
\alias{upload_attachments}
\title{Upload NocoDB attachments}
\usage{
upload_attachments(
  paths,
  names = fs::path_file(paths),
  types = mime::guess_type(paths),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg),
  max_tries = 5L,
  verbosity = NULL
)
}
\arguments{
\item{paths}{Path(s) to the file(s) to be uploaded. A character vector.}

\item{names}{File name(s) to assign to the uploaded file(s). A character vector. Note that a random string like \verb{_NnW3C} is always appended to the resulting
file name before its file type suffix, so \code{name = "some-pic.jpg"} will result in something like \code{some-pic_NnW3C.jpg}, for example.}

\item{types}{\href{https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types}{Media type}(s) of the uploaded files. A character vector.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}

\item{max_tries}{Maximum number of attempts to retry in case of an HTTP error. An integerish scalar.}

\item{verbosity}{How much information to print? This is a wrapper
around \code{req_verbose()} that uses an integer to control verbosity:
\itemize{
\item 0: no output
\item 1: show headers
\item 2: show headers and bodies
\item 3: show headers, bodies, and curl status messages.
}

Use \code{\link[httr2:with_verbosity]{with_verbosity()}} to control the verbosity of requests that
you can't affect directly.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} containing metadata about the uploaded attachments.
}
\description{
Uploads the specified files to a NocoDB server as \href{https://docs.nocodb.com/fields/field-types/custom-types/attachment/}{attachments} via its
\href{https://meta-apis-v2.nocodb.com/#tag/Storage}{\code{POST /api/v2/storage/upload}} API endpoint. Uploaded files are stored under \verb{/nc/uploads/} in the configured
object storage bucket (if one is set up) or directly on the server's filesystem.
}
\concept{attachments}
