% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{users}
\alias{users}
\title{Get NocoDB users metadata}
\usage{
users(
  base_id = base_id(),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  auth_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)
}
\arguments{
\item{base_id}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{auth_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API authentication token}. A character scalar.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with metadata about the users in the specified base on a NocoDB server from its
\href{https://meta-apis-v2.nocodb.com/#tag/Auth/operation/auth-base-user-list}{\code{GET /api/v2/meta/bases/{base_id}/users}} API endpoint.
}
\seealso{
Other functions to manage NocoDB users:
\code{\link{create_user}()}
}
\concept{users}
