% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{validate_user_email}
\alias{validate_user_email}
\title{Validate NocoDB user e-mail}
\usage{
validate_user_email(
  verification_token,
  origin = pal::pkg_config_val("origin"),
  quiet = FALSE
)
}
\arguments{
\item{verification_token}{User's e-mail verification token.}

\item{origin}{NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}. A character scalar.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Validates a user's e-mail address on a NocoDB server via its
\href{https://data-apis-v1.nocodb.com/#tag/Auth/operation/auth-email-validate}{\code{POST /api/v1/auth/email/validate/{verification_token}}} API endpoint.
}
\details{
You can use \code{\link[=update_user]{update_user()}} to get a tibble containing a user's \code{email_verification_token}.
}
\seealso{
Other functions to manage NocoDB users:
\code{\link{add_user}()},
\code{\link{base_users}()},
\code{\link{delete_base_user}()},
\code{\link{delete_user}()},
\code{\link{invite_base_user}()},
\code{\link{invite_user}()},
\code{\link{resend_base_user_invitation}()},
\code{\link{sign_up_user}()},
\code{\link{update_base_user}()},
\code{\link{update_user}()},
\code{\link{user_id}()},
\code{\link{users}()},
\code{\link{whoami}()}
}
\concept{users}
