% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{access_token}
\alias{access_token}
\title{Get user's current access token}
\usage{
access_token(
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg, require = TRUE)
)
}
\arguments{
\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}
}
\value{
Access token as a character scalar. \code{NA_character_} if no access token exists for \code{hostname} and \code{email}.
}
\description{
Retrieves the access token belonging to the user with the specified \code{email} from the corresponding package environment. An access token is only available if
the user has \link[=sign_in]{signed in} or \link[=refresh_sign_in]{refreshed their access token} during the current \R session.
}
\seealso{
Other access_token: 
\code{\link{decode_access_token}()},
\code{\link{is_access_token_expired}()},
\code{\link{store_access_token}()}
}
\concept{access_token}
\keyword{internal}
