% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{add_user}
\alias{add_user}
\title{Add NocoDB user account}
\usage{
add_user(
  user_email,
  user_password,
  display_name = NULL,
  subscribe_to_newsletter = FALSE,
  origin = funky::config_val("origin"),
  email = funky::config_val("email", require = TRUE),
  password = funky::config_val("password", require = TRUE),
  quiet = TRUE
)
}
\arguments{
\item{user_email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the user to sign up.}

\item{user_password}{\code{\link[base:character]{character(1)}}\cr
Password of the user to sign up.}

\item{display_name}{\code{\link[base:character]{character(1)}}\cr
Name to be displayed for the user in NocoDB. Omitted if \code{NULL}.}

\item{subscribe_to_newsletter}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to subscribe the signed up user to the NocoDB newsletter.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{quiet}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to suppress printing status output from internal processing.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with metadata about the updated NocoDB user, invisibly.
}
\description{
Adds a NocoDB user account on a NocoDB server without any user interaction. E-mail invitation is bypassed and the user's e-mail address is automatically
\link[=validate_user_email]{validated}. A convenience function that combines several of the other functions found in this package.
}
\details{
\code{add_user()} does not support authentication via \href{https://docs.nocodb.com/account-settings/api-tokens/}{API tokens} because of \code{\link[=update_app_settings]{update_app_settings()}}.
}
\seealso{
Other functions to manage NocoDB users:
\code{\link{base_users}()},
\code{\link{delete_base_user}()},
\code{\link{delete_user}()},
\code{\link{invite_base_user}()},
\code{\link{invite_user}()},
\code{\link{resend_base_user_invitation}()},
\code{\link{sign_up_user}()},
\code{\link{update_base_user}()},
\code{\link{update_user}()},
\code{\link{user_id}()},
\code{\link{users}()},
\code{\link{validate_user_email}()},
\code{\link{whoami}()}
}
\concept{users}
