% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{add_user}
\alias{add_user}
\title{Add NocoDB user account}
\usage{
add_user(
  email_new,
  password_new,
  display_name = NULL,
  subscribe_to_newsletter = FALSE,
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg, required = FALSE),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg, required = FALSE),
  quiet = TRUE
)
}
\arguments{
\item{email_new}{E-mail address of the user to sign up. A character scalar.}

\item{password_new}{Password of the user to sign up. A character scalar.}

\item{display_name}{Name to be displayed for the user in NocoDB.}

\item{subscribe_to_newsletter}{Whether or not to subscribe the signed up user to the NocoDB newsletter.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Adds a NocoDB user account on a NocoDB server without any user interaction. E-mail invitation is bypassed and the user's e-mail address is automatically
\link[=validate_user_email]{validated}. A convenience function that combines several of the other functions found in this package.
}
\details{
\code{add_user()} does not support authentication via \href{https://docs.nocodb.com/account-settings/api-tokens/}{API tokens} because of \code{\link[=update_app_settings]{update_app_settings()}}.
}
\seealso{
Other functions to manage NocoDB users:
\code{\link{base_users}()},
\code{\link{invite_base_user}()},
\code{\link{invite_user}()},
\code{\link{sign_up_user}()},
\code{\link{update_base_user}()},
\code{\link{update_user}()},
\code{\link{user}()},
\code{\link{user_id}()},
\code{\link{users}()},
\code{\link{validate_user_email}()}
}
\concept{users}
