% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{api}
\alias{api}
\title{Call NocoDB API}
\usage{
api(
  path,
  method = c("GET", "CONNECT", "DELETE", "HEAD", "OPTIONS", "PATCH", "POST", "PUT",
    "TRACE"),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg),
  auth = TRUE,
  body_json = NULL,
  auto_unbox = TRUE,
  simplify = TRUE,
  flatten = TRUE,
  max_tries = 3L,
  verbosity = NULL
)
}
\arguments{
\item{path}{Optionally, path to save body of the response. This is useful
for large responses since it avoids storing the response in memory.}

\item{method}{\href{https://developer.mozilla.org/en-US/docs/Web/HTTP/Methods}{HTTP request method}. One of
\code{"GET"}, \code{"CONNECT"}, \code{"DELETE"}, \code{"HEAD"}, \code{"OPTIONS"}, \code{"PATCH"}, \code{"POST"}, \code{"PUT"} or \code{"TRACE"}.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}

\item{auth}{Whether or not to include an \link[=req_auth]{authentication header} in the HTTP request.}

\item{body_json}{Data to include as JSON in the HTTP request body. Either a list or \code{NULL} for an empty body.}

\item{auto_unbox}{Whether or not to automatically "unbox" length-1 vectors in \code{body_json} to JSON scalars.}

\item{simplify}{Whether or not to automatically simplify JSON structures in the returned JSON. Enables/disables all \verb{simplify*} arguments of
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}.}

\item{flatten}{Whether or not to automatically \link[jsonlite:flatten]{flatten} nested data frames in the returned JSON into a single non-nested data frame.}

\item{max_tries}{Maximum number of attempts to retry in case of an HTTP error. An integerish scalar.}

\item{verbosity}{How much information to print? This is a wrapper
around \code{req_verbose()} that uses an integer to control verbosity:
\itemize{
\item 0: no output
\item 1: show headers
\item 2: show headers and bodies
\item 3: show headers, bodies, and curl status messages.
}

Use \code{\link[httr2:with_verbosity]{with_verbosity()}} to control the verbosity of requests that
you can't affect directly.}
}
\value{
A list if the response body is of type JSON, otherwise a character scalar.
}
\description{
Performs an API call to a NocoDB server and returns the response as a list if it is of type JSON, otherwise as a character scalar.
}
\details{
See \code{\link[=req_auth]{req_auth()}} for details about the authentication logic enabled via \code{auth}.

For full control over the performed HTTP request, resort to the lower-level \code{\link[=req_basic]{req_basic()}}.
}
\seealso{
Other common functions:
\code{\link{req_auth}()},
\code{\link{req_basic}()}
}
\concept{common}
