% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{assert_super_admin}
\alias{assert_super_admin}
\title{Assert super admin}
\usage{
assert_super_admin(
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = NULL
)
}
\arguments{
\item{origin}{NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Asserts that the provided credentials belong to NocoDB's \href{https://docs.nocodb.com/collaboration/oss-specific-details/#base-level-permissions}{super admin user}.
}
\details{
The user is determined based on \code{api_token} or \code{email} and \code{password} (the former takes precedence).
}
\seealso{
Other functions to manage NocoDB authentication:
\code{\link{api_tokens}()},
\code{\link{create_api_token}()},
\code{\link{delete_api_token}()},
\code{\link{is_signed_in}()},
\code{\link{is_super_admin}()},
\code{\link{refresh_sign_in}()},
\code{\link{req_auth}()},
\code{\link{sign_in}()},
\code{\link{sign_out}()},
\code{\link{sign_up_user}()}

Other user: 
\code{\link{is_signed_in}()},
\code{\link{is_super_admin}()},
\code{\link{refresh_sign_in}()},
\code{\link{sign_in}()},
\code{\link{sign_out}()}
}
\concept{auth}
\concept{user}
\keyword{internal}
