% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{base}
\alias{base}
\title{Get NocoDB base}
\usage{
base(
  id_base = base_id(origin = origin, email = email, password = password, api_token =
    api_token),
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token")
)
}
\arguments{
\item{id_base}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{origin}{NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with metadata about the specified NocoDB base.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with metadata about the specified base on a NocoDB server from its
\href{https://meta-apis-v2.nocodb.com/#tag/Base/operation/base-read}{\code{GET /api/v2/meta/bases/{id_base}}} API endpoint.
}
\seealso{
Other functions to manage NocoDB bases:
\code{\link{base_id}()},
\code{\link{base_ui_acl}()},
\code{\link{base_users}()},
\code{\link{bases}()},
\code{\link{create_base}()},
\code{\link{delete_base}()},
\code{\link{delete_base_user}()},
\code{\link{invite_base_user}()},
\code{\link{resend_base_user_invitation}()},
\code{\link{update_base}()},
\code{\link{update_base_ui_acl}()},
\code{\link{update_base_user}()}
}
\concept{bases}
