% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{bases}
\alias{bases}
\title{List NocoDB bases}
\usage{
bases(
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token")
)
}
\arguments{
\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with metadata about all bases on a NocoDB server from its
\href{https://meta-apis-v2.nocodb.com/#tag/Base/operation/base-list}{\code{GET /api/v2/meta/bases}} API endpoint.
}
\seealso{
Other functions to manage NocoDB bases:
\code{\link{base}()},
\code{\link{base_id}()},
\code{\link{base_ui_acl}()},
\code{\link{base_users}()},
\code{\link{create_base}()},
\code{\link{delete_base}()},
\code{\link{delete_base_user}()},
\code{\link{invite_base_user}()},
\code{\link{resend_base_user_invitation}()},
\code{\link{update_base}()},
\code{\link{update_base_ui_acl}()},
\code{\link{update_base_user}()}
}
\concept{bases}
