% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{create_api_token}
\alias{create_api_token}
\title{Create NocoDB API token}
\usage{
create_api_token(
  description,
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email", require = TRUE),
  password = pal::pkg_config_val("password", require = TRUE)
)
}
\arguments{
\item{description}{\code{\link[base:character]{character(1)}}\cr
Description to assign to the newly created API token (displayed as \emph{Token name} in NocoDB's UI).}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with (meta)data about the newly created API token.
}
\description{
Adds a new \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token} for the specified user on a NocoDB server via its
\href{https://docs.nocodb.com/0.109.7/developer-resources/rest-apis/#meta-apis}{\code{POST /api/v1/tokens}} API endpoint.
}
\details{
This API endpoint does not support authentication via \href{https://docs.nocodb.com/account-settings/api-tokens/}{API tokens}.
}
\seealso{
Other functions to manage NocoDB API tokens:
\code{\link{api_tokens}()},
\code{\link{delete_api_token}()}

Other functions to manage NocoDB authentication:
\code{\link{api_tokens}()},
\code{\link{assert_super_admin}()},
\code{\link{delete_api_token}()},
\code{\link{is_signed_in}()},
\code{\link{is_super_admin}()},
\code{\link{refresh_sign_in}()},
\code{\link{req_auth}()},
\code{\link{sign_in}()},
\code{\link{sign_out}()},
\code{\link{sign_up_user}()}
}
\concept{api_tokens}
\concept{auth}
