% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{create_base}
\alias{create_base}
\title{Create NocoDB base}
\usage{
create_base(
  title = funky::config_val("base_title"),
  description = NULL,
  color = "#36BFFF",
  show_null_and_empty_in_filter = TRUE,
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)
}
\arguments{
\item{title}{\code{\link[base:character]{character(1)}}\cr
Base title.}

\item{description}{\code{\link[base:character]{character(1)}}\cr
Base description.}

\item{color}{\code{\link[base:character]{character(1)}}\cr
Color to use for the base icon. A hexadecimal color code like "#36BFFF".}

\item{show_null_and_empty_in_filter}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to distinguish \code{NULL}s from empty values in column filters. If \code{FALSE}, \code{NULL} and empty fields are
treated alike.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with metadata about the newly created NocoDB base, invisibly.
}
\description{
Adds a new base on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/Bases/operation/base-create}{\code{POST /api/v2/meta/bases}} API endpoint.
}
\seealso{
Other functions to manage NocoDB bases:
\code{\link{base}()},
\code{\link{base_id}()},
\code{\link{base_ui_acl}()},
\code{\link{base_users}()},
\code{\link{bases}()},
\code{\link{delete_base}()},
\code{\link{delete_base_user}()},
\code{\link{invite_base_user}()},
\code{\link{resend_base_user_invitation}()},
\code{\link{update_base}()},
\code{\link{update_base_ui_acl}()},
\code{\link{update_base_user}()}
}
\concept{bases}
