% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{create_data_src}
\alias{create_data_src}
\title{Create NocoDB data source}
\usage{
create_data_src(
  connection,
  id_integration = NULL,
  type = c("mssql", "mysql", "pg", "sqlite3"),
  alias = NULL,
  inflection_column = c("none", "camelize"),
  inflection_table = c("none", "camelize"),
  is_schema_readonly = FALSE,
  is_data_readonly = FALSE,
  enabled = TRUE,
  id_base = base_id(origin = origin, email = email, password = password, api_token =
    api_token),
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token")
)
}
\arguments{
\item{connection}{Type-specific connection details for the data source. A list.}

\item{id_integration}{\code{\link[base:character]{character(1)}}\cr
Optional identifier of an existing NocoDB database integration as returned by \code{\link[=integration_id]{integration_id()}}, of which to reuse connection details}

\item{type}{Type of the data source. One of \code{"mssql"}, \code{"mysql"}, \code{"pg"} or \code{"sqlite3"}.}

\item{alias}{Alias of the data source. A character scalar.}

\item{inflection_column}{Type of inflection to apply for column names in the data source. One of
\code{"none"} or \code{"camelize"}.}

\item{inflection_table}{Type of inflection to apply for table names in the data source. One of
\code{"none"} or \code{"camelize"}.}

\item{is_schema_readonly}{Whether or not to \href{https://docs.nocodb.com/data-sources/connect-to-data-source/#allow-schema-edit}{forbid alterations of the data source's \strong{schema} from NocoDB}.}

\item{is_data_readonly}{Whether or not to \href{https://docs.nocodb.com/data-sources/connect-to-data-source/#allow-data-edit}{forbid alterations of the data source's \strong{data} from NocoDB}.}

\item{enabled}{Whether the added data source is enabled or disabled.}

\item{id_base}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{origin}{NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
\code{alias}, invisibly.
}
\description{
Adds a data source to the specified base on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/Source/operation/source-create}{\code{POST /api/v2/meta/bases/{id_base}/sources}} API endpoint.
}
\examples{
\dontrun{
nocodb::create_data_src(type = "pg",
                        connection = list(host = "REPLACE-ME",
                                          port = 5432,
                                          ssl = list(ca = "",
                                                     cert = "",
                                                     key = ""),
                                          sslmode = "verify-full",
                                          user = "REPLACE-ME",
                                          password = "REPLACE-ME",
                                          database = "REPLACE-ME"),
                        alias = "REPLACE-ME",
                        inflection_column = "none",
                        inflection_table = "none")}
}
\seealso{
Other functions to manage NocoDB base data sources:
\code{\link{create_data_src_tbl}()},
\code{\link{data_src}()},
\code{\link{data_src_diff}()},
\code{\link{data_src_id}()},
\code{\link{data_src_tbls}()},
\code{\link{data_srcs}()},
\code{\link{delete_data_src}()},
\code{\link{has_data_src_diff}()},
\code{\link{sync_data_src}()},
\code{\link{test_data_src}()},
\code{\link{update_data_src}()}
}
\concept{data_src}
