% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{data_src_id}
\alias{data_src_id}
\title{Get NocoDB data source ID}
\usage{
data_src_id(
  alias,
  id_base = base_id(hostname = hostname),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)
}
\arguments{
\item{alias}{Alias of the data source. A character scalar.}

\item{id_base}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
Data source identifier as a character scalar.
}
\description{
Returns the identifier of the data source with the specified alias in the specified base on a NocoDB server.
}
\seealso{
Other functions to manage NocoDB base data sources:
\code{\link{create_data_src}()},
\code{\link{data_src}()},
\code{\link{data_srcs}()},
\code{\link{delete_data_src}()}
}
\concept{data_src}
