% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{data_src_tbls}
\alias{data_src_tbls}
\title{List tables in NocoDB data source}
\usage{
data_src_tbls(
  id_data_src,
  id_base = base_id(hostname = hostname, email = email, password = password, api_token =
    api_token),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)
}
\arguments{
\item{id_data_src}{NocoDB data source identifier as returned by \code{\link[=data_src_id]{data_src_id()}}. A character scalar.}

\item{id_base}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with metadata about all tables in the specified data source of the specified base on a NocoDB server from its
\href{https://meta-apis-v2.nocodb.com/#tag/Source/operation/table-list}{\code{GET /api/v2/meta/bases/{id_base}/{id_data_src}/tables}} API endpoint.
}
\seealso{
Other functions to manage NocoDB base data sources:
\code{\link{create_data_src}()},
\code{\link{create_data_src_tbl}()},
\code{\link{data_src}()},
\code{\link{data_src_diff}()},
\code{\link{data_src_id}()},
\code{\link{data_srcs}()},
\code{\link{delete_data_src}()},
\code{\link{has_data_src_diff}()},
\code{\link{sync_data_src}()},
\code{\link{test_data_src}()},
\code{\link{update_data_src}()}

Other functions to manage NocoDB tables:
\code{\link{create_data_src_tbl}()},
\code{\link{create_tbl}()},
\code{\link{delete_tbl}()},
\code{\link{reorder_tbl}()},
\code{\link{set_tbl_metadata}()},
\code{\link{tbl}},
\code{\link{tbl_id}()},
\code{\link{tbls}()},
\code{\link{update_tbl}()}
}
\concept{data_src}
\concept{tbls}
