% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{decode_access_token}
\alias{decode_access_token}
\title{Decode access token (JWT)}
\usage{
decode_access_token(x)
}
\arguments{
\item{x}{The access token to be decoded.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Decodes an access token that adheres to the \href{https://de.wikipedia.org/wiki/JSON_Web_Token}{JSON Web Token (JWT)} standard and returns it as a
\link[tibble:tbl_df-class]{tibble}. Only the payload is returned.
}
\keyword{internal}
