% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{delete_api_token}
\alias{delete_api_token}
\title{Delete NocoDB API token}
\usage{
delete_api_token(
  api_token,
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg, require = TRUE),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg, require = TRUE)
)
}
\arguments{
\item{api_token}{NocoDB API token to delete.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}
}
\value{
\code{api_token}, invisibly.
}
\description{
Deletes the specified API token on a NocoDB server via its
\href{https://docs.nocodb.com/0.109.7/developer-resources/rest-apis/#meta-apis}{\code{DELETE /api/v1/tokens/{api_token}}} API endpoint.
}
\details{
This API endpoint does not support authentication via \href{https://docs.nocodb.com/account-settings/api-tokens/}{API tokens}.
}
\seealso{
Other functions to manage NocoDB API tokens:
\code{\link{api_tokens}()},
\code{\link{create_api_token}()}

Other functions to manage NocoDB authentication:
\code{\link{api_tokens}()},
\code{\link{assert_super_admin}()},
\code{\link{create_api_token}()},
\code{\link{is_signed_in}()},
\code{\link{is_super_admin}()},
\code{\link{refresh_sign_in}()},
\code{\link{req_auth}()},
\code{\link{sign_in}()},
\code{\link{sign_out}()},
\code{\link{sign_up_user}()}
}
\concept{api_tokens}
\concept{auth}
