% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{delete_tbl_col}
\alias{delete_tbl_col}
\title{Delete NocoDB table column}
\usage{
delete_tbl_col(
  id_col,
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)
}
\arguments{
\item{id_col}{\code{\link[base:character]{character(1)}}\cr
NocoDB column identifier as returned by \code{\link[=tbl_col_id]{tbl_col_id()}}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
\code{id_col}, invisibly.
}
\description{
Deletes the specified table column on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/Fields/operation/db-table-column-delete}{\code{DELETE /api/v2/meta/columns/{id_col}}} API endpoint.
}
\seealso{
Other functions to manage NocoDB table columns:
\code{\link{create_tbl_col}()},
\code{\link{set_display_val}()},
\code{\link{set_display_vals}()},
\code{\link{tbl_col}()},
\code{\link{tbl_col_id}()},
\code{\link{tbl_cols}()},
\code{\link{update_tbl_col}()}
}
\concept{cols}
