% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{delete_user}
\alias{delete_user}
\title{Delete NocoDB user}
\usage{
delete_user(
  id_user,
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg),
  quiet = FALSE
)
}
\arguments{
\item{id_user}{NocoDB user identifier as returned by \code{\link[=user_id]{user_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{id_user}, invisibly.
}
\description{
Deletes the specified user on a NocoDB server via its
\href{https://docs.nocodb.com/0.109.7/developer-resources/rest-apis/#meta-apis}{\code{DELETE /api/v1/users/{id_user}}} API endpoint.
}
\details{
This API endpoint does not support authentication via \href{https://docs.nocodb.com/account-settings/api-tokens/}{API tokens}.
}
\section{API errors}{
\itemize{
\item If you encounter a \strong{\verb{Cannot read properties of undefined (reading 'roles')}} error, it likely means the specified \code{id_user} is invalid (i.e. doesn't
exist).
\item If you encounter a \strong{\verb{userDelete -  : Not allowed}} error, it means the authenticating user doesn't have sufficient privileges (must be
\code{"org-level-creator"}).
}
}

\seealso{
Other functions to manage NocoDB users:
\code{\link{add_user}()},
\code{\link{base_users}()},
\code{\link{delete_base_user}()},
\code{\link{invite_base_user}()},
\code{\link{invite_user}()},
\code{\link{resend_base_user_invitation}()},
\code{\link{sign_up_user}()},
\code{\link{update_base_user}()},
\code{\link{update_user}()},
\code{\link{user_id}()},
\code{\link{users}()},
\code{\link{validate_user_email}()},
\code{\link{whoami}()}
}
\concept{users}
