% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{integration}
\alias{integration}
\title{Get NocoDB integration}
\usage{
integration(
  id_integration,
  incl_config = TRUE,
  hostname = pal::pkg_config_val("hostname"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token")
)
}
\arguments{
\item{id_integration}{\code{\link[base:character]{character(1)}}\cr
NocoDB integration identifier as returned by \code{\link[=integration_id]{integration_id()}}.}

\item{incl_config}{Whether or not to include the integration configuration in the result (as a separate \code{config} column).}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with metadata about the specified integration on a NocoDB server from its \code{GET /api/v2/meta/integrations/{id_integration}}
API endpoint.
}
\seealso{
Other functions to manage NocoDB integrations:
\code{\link{create_integration}()},
\code{\link{delete_integration}()},
\code{\link{integration_id}()},
\code{\link{integrations}()},
\code{\link{update_integration}()}
}
\concept{integrations}
