% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{invite_user}
\alias{invite_user}
\title{Invite NocoDB user}
\usage{
invite_user(
  user_email,
  org_role = c("org-level-viewer", "org-level-creator"),
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email", require = TRUE),
  password = pal::pkg_config_val("password", require = TRUE),
  quiet = FALSE
)
}
\arguments{
\item{user_email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the user to invite.}

\item{org_role}{\code{\link[base:character]{character(1)}}\cr
\href{https://docs.nocodb.com/collaboration/oss-specific-details/#organisation-level-permissions}{Organisation-level role} to assign to the user. One of
\code{"org-level-viewer"} or \code{"org-level-creator"}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{quiet}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Invites a new user on a NocoDB server via its
\href{https://docs.nocodb.com/0.109.7/developer-resources/rest-apis/#meta-apis}{\code{POST /api/v1/users}} API endpoint.
}
\details{
This API endpoint does not support authentication via \href{https://docs.nocodb.com/account-settings/api-tokens/}{API tokens}.
}
\seealso{
Other functions to manage NocoDB users:
\code{\link{add_user}()},
\code{\link{base_users}()},
\code{\link{delete_base_user}()},
\code{\link{delete_user}()},
\code{\link{invite_base_user}()},
\code{\link{resend_base_user_invitation}()},
\code{\link{sign_up_user}()},
\code{\link{update_base_user}()},
\code{\link{update_user}()},
\code{\link{user_id}()},
\code{\link{users}()},
\code{\link{validate_user_email}()},
\code{\link{whoami}()}
}
\concept{users}
