% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{is_plugin_active}
\alias{is_plugin_active}
\title{Test if plugin is active}
\usage{
is_plugin_active(
  title,
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)
}
\arguments{
\item{title}{NocoDB plugin title. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
A logical scalar.
}
\description{
Tests whether the plugin with the specified \code{title} is active on a NocoDB server via its
\href{https://docs.nocodb.com/0.109.7/developer-resources/rest-apis/#meta-apis}{\code{GET /api/v1/db/meta/plugins/{title}/status}} API endpoint.
}
\details{
Note that the API endpoint doesn't check the plugin \code{title} for validity and simply returns \code{FALSE} for a \code{title} which does not actually exist.
}
\seealso{
Other functions to manage NocoDB plugins:
\code{\link{plugin}()},
\code{\link{plugin_category}()},
\code{\link{plugin_id}()},
\code{\link{plugins}()},
\code{\link{test_plugin}()},
\code{\link{update_plugin}()}
}
\concept{plugins}
