% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{is_signed_in}
\alias{is_signed_in}
\title{Test if user is signed in}
\usage{
is_signed_in(
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg, require = TRUE)
)
}
\arguments{
\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}
}
\value{
A logical scalar.
}
\description{
Tests whether the user with the specified \code{email} has signed in during the current \R session. See \code{\link[=sign_in]{sign_in()}} for details.
}
\seealso{
Other functions to manage NocoDB authentication:
\code{\link{api_tokens}()},
\code{\link{assert_super_admin}()},
\code{\link{create_api_token}()},
\code{\link{delete_api_token}()},
\code{\link{is_super_admin}()},
\code{\link{refresh_sign_in}()},
\code{\link{req_auth}()},
\code{\link{sign_in}()},
\code{\link{sign_out}()},
\code{\link{sign_up_user}()}

Other user: 
\code{\link{assert_super_admin}()},
\code{\link{is_super_admin}()},
\code{\link{refresh_sign_in}()},
\code{\link{sign_in}()},
\code{\link{sign_out}()}
}
\concept{auth}
\concept{user}
\keyword{internal}
