% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb-package.R
\docType{package}
\name{nocodb-package}
\alias{nocodb}
\alias{nocodb-package}
\title{nocodb: Use NocoDB's RESTful APIs}
\description{
nocodb allows to work with NocoDB's RESTful APIs from R. NocoDB is a web application that provides an intuitive spreadsheet-like front-end for relational databases, see https://nocodb.com.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/rpkg.dev/nocodb}
  \item Report bugs at \url{https://gitlab.com/rpkg.dev/nocodb/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
