% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{refresh_sign_in}
\alias{refresh_sign_in}
\title{Refresh NocoDB access token}
\usage{
refresh_sign_in(
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email", require = TRUE)
)
}
\arguments{
\item{origin}{NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}. A character scalar.}

\item{email}{E-mail address of the NocoDB user whose access token is to be refreshed. A character scalar.}
}
\value{
The freshly generated access token as a character scalar, invisibly.
}
\description{
Requests a fresh NocoDB access token via the \href{https://data-apis-v1.nocodb.com/#tag/Auth/operation/auth-token-refresh}{\code{POST /api/v1/auth/token/refresh}} API
endpoint. Requires a valid refresh token cached to disk by \code{\link[=sign_in]{sign_in(cache_refresh_token = TRUE)}}.
}
\details{
Besides generating a fresh access token, also a new refresh token is cached to disk.
}
\seealso{
Other functions to manage NocoDB authentication:
\code{\link{api_tokens}()},
\code{\link{assert_super_admin}()},
\code{\link{create_api_token}()},
\code{\link{delete_api_token}()},
\code{\link{is_signed_in}()},
\code{\link{is_super_admin}()},
\code{\link{req_auth}()},
\code{\link{sign_in}()},
\code{\link{sign_out}()},
\code{\link{sign_up_user}()}

Other user: 
\code{\link{assert_super_admin}()},
\code{\link{is_signed_in}()},
\code{\link{is_super_admin}()},
\code{\link{sign_in}()},
\code{\link{sign_out}()}
}
\concept{auth}
\concept{user}
\keyword{internal}
