% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{reorder_tbl}
\alias{reorder_tbl}
\title{Re-order NocoDB table}
\usage{
reorder_tbl(
  tbl_id = tbl_id(),
  order = 1L,
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  auth_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)
}
\arguments{
\item{tbl_id}{NocoDB table identifier as returned by \code{\link[=tbl_id]{tbl_id()}}. A character scalar.}

\item{order}{A number to assign as the table's order "weight".}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{auth_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API authentication token}. A character scalar.}
}
\value{
\code{tbl_id}, invisibly.
}
\description{
Sets the numeric order of the specified table on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/DB-Table/operation/db-table-reorder}{\code{POST /api/v2/meta/tables/{tbl_id}/reorder}} API endpoint.
}
\details{
Lower numbers place the table higher up in the UI and vice versa. The current order of all the tables in a base can be determined via \code{\link[=tbls]{tbls()}}.
}
\seealso{
Other functions to manage NocoDB tables:
\code{\link{tbl}},
\code{\link{tbl_id}()},
\code{\link{tbls}()},
\code{\link{update_tbl}()}
}
\concept{tbls}
