% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{reorder_tbl}
\alias{reorder_tbl}
\title{Re-order NocoDB table}
\usage{
reorder_tbl(
  id_tbl,
  order = 1L,
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)
}
\arguments{
\item{id_tbl}{NocoDB table identifier as returned by \code{\link[=tbl_id]{tbl_id()}}. A character scalar.}

\item{order}{A number to assign as the table's NocoDB-specific order "weight".}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
\code{id_tbl}, invisibly.
}
\description{
Sets the numeric order of the specified table on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/DB-Table/operation/db-table-reorder}{\code{POST /api/v2/meta/tables/{id_tbl}/reorder}} API endpoint.
}
\details{
Lower numbers place the table higher up in the UI and vice versa. The current order of all the tables in a base can be determined via \code{\link[=tbls]{tbls()}}.
}
\seealso{
Other functions to manage NocoDB tables:
\code{\link{create_data_src_tbl}()},
\code{\link{create_tbl}()},
\code{\link{data_src_tbls}()},
\code{\link{delete_tbl}()},
\code{\link{set_tbl_metadata}()},
\code{\link{tbl}},
\code{\link{tbl_id}()},
\code{\link{tbls}()},
\code{\link{update_tbl}()}
}
\concept{tbls}
