% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{set_display_vals}
\alias{set_display_vals}
\title{Set all NocoDB display value columns}
\usage{
set_display_vals(
  data,
  base_id = base_id(),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  auth_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg),
  quiet = FALSE
)
}
\arguments{
\item{data}{Dataframe with the two columns \code{name} and \code{display_col} to source the . Other columns are ignored.}

\item{base_id}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{auth_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API authentication token}. A character scalar.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Sets the proper column as the \href{https://docs.nocodb.com/fields/display-value/}{display value} for all tables in \code{data}. A convenience function that combines
several of the other functions found in this package.
}
\seealso{
Other functions to manage NocoDB table columns:
\code{\link{set_display_val}()},
\code{\link{tbl_col}()},
\code{\link{tbl_col_id}()},
\code{\link{tbl_cols}()},
\code{\link{update_tbl_col}()}
}
\concept{cols}
