% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{sign_out}
\alias{sign_out}
\title{Sign out NocoDB user}
\usage{
sign_out(
  origin = funky::config_val("origin"),
  email = funky::config_val("email", require = TRUE),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token"),
  quiet = FALSE
)
}
\arguments{
\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to sign out.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}

\item{quiet}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{email}, invisibly.
}
\description{
Invalidates a cached refresh token via the \href{https://data-apis-v1.nocodb.com/#tag/Auth/operation/auth-signout}{\code{POST /api/v1/auth/user/signout}} API endpoint
and clears the associated cache on disk.
}
\details{
Note that the \code{email} address to sign out and the \code{api_token} to authenticate do not necessarily need to belong to the same user.
}
\seealso{
Other functions to manage NocoDB authentication:
\code{\link{api_tokens}()},
\code{\link{assert_super_admin}()},
\code{\link{create_api_token}()},
\code{\link{delete_api_token}()},
\code{\link{is_signed_in}()},
\code{\link{is_super_admin}()},
\code{\link{refresh_sign_in}()},
\code{\link{req_auth}()},
\code{\link{sign_in}()},
\code{\link{sign_up_user}()}

Other user: 
\code{\link{assert_super_admin}()},
\code{\link{is_signed_in}()},
\code{\link{is_super_admin}()},
\code{\link{refresh_sign_in}()},
\code{\link{sign_in}()}
}
\concept{auth}
\concept{user}
