% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{sign_out}
\alias{sign_out}
\title{Sign out NocoDB user}
\usage{
sign_out(
  email = pal::pkg_config_val(key = "email", pkg = this_pkg),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg, required = FALSE),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg, required = FALSE),
  quiet = FALSE
)
}
\arguments{
\item{email}{E-mail address of the NocoDB user to sign out. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{email}, invisibly.
}
\description{
Invalidates a cached refresh token via the \href{https://data-apis-v1.nocodb.com/#tag/Auth/operation/auth-signout}{\code{POST /api/v1/auth/user/signout}} API endpoint
and clears the associated cache on disk.
}
\details{
Note that the \code{email} address to sign out and the \code{api_token} to authenticate do not necessarily need to belong to the same user.
}
\seealso{
Other functions to manage NocoDB authentication:
\code{\link{is_signed_in}()},
\code{\link{refresh_sign_in}()},
\code{\link{req_auth}()},
\code{\link{sign_in}()},
\code{\link{sign_up_user}()}
}
\concept{auth}
