% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{store_access_token}
\alias{store_access_token}
\title{Store NocoDB access token}
\usage{
store_access_token(
  x,
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email", require = TRUE)
)
}
\arguments{
\item{x}{\code{\link[base:character]{character(1)}}\cr
Access token to be stored.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}
}
\value{
\code{x}, invisibly.
}
\description{
Stores a NocoDB user's access token in the \code{stateful} package environment.
}
\seealso{
Other access_token: 
\code{\link{access_token}()},
\code{\link{decode_access_token}()},
\code{\link{is_access_token_expired}()}
}
\concept{access_token}
\keyword{internal}
