% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{sync_data_src_eagerly}
\alias{sync_data_src_eagerly}
\title{Synchronize NocoDB data source schema eagerly}
\usage{
sync_data_src_eagerly(
  id_data_src,
  id_base = base_id(hostname = hostname, email = email, password = password, api_token =
    api_token),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg),
  wait_max = 30L,
  wait_resync = 7
)
}
\arguments{
\item{id_data_src}{NocoDB data source identifier as returned by \code{\link[=data_src_id]{data_src_id()}}. A character scalar.}

\item{id_base}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}

\item{wait_max}{Maximum time in seconds to wait for schema synchronization to finish. An error is thrown if there are still schema changes pending when the
timeout is reached. An integerish scalar \verb{>= wait_resync}.}

\item{wait_resync}{Interval in seconds to repeat calling \code{\link[=sync_data_src]{sync_data_src()}}. A number \verb{>= 1} and \verb{<= wait_max}.}
}
\value{
\code{id_data_src}, invisibly.
}
\description{
Synchronizes the specified data source's schema between its external and its NocoDB-internal state by repeatedly calling \code{\link[=sync_data_src]{sync_data_src()}} until
\code{\link[=has_data_src_diff]{has_data_src_diff()}} doesn't detect any more changes.
}
\concept{admin}
\concept{nocodb}
