% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{tbl_id}
\alias{tbl_id}
\title{Get NocoDB table ID}
\usage{
tbl_id(
  name,
  id_base = base_id(origin = origin, email = email, password = password, api_token =
    api_token),
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token")
)
}
\arguments{
\item{name}{\code{\link[base:character]{character(1)}}\cr
NocoDB table name.}

\item{id_base}{\code{\link[base:character]{character(1)}}\cr
NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
A character scalar.
}
\description{
Returns the identifier of the table with the specified name in the specified base on a NocoDB server.
}
\seealso{
Other functions to manage NocoDB tables:
\code{\link{create_data_src_tbl}()},
\code{\link{create_tbl}()},
\code{\link{data_src_tbls}()},
\code{\link{delete_tbl}()},
\code{\link{reorder_tbl}()},
\code{\link{set_tbl_metadata}()},
\code{\link{tbl}},
\code{\link{tbls}()},
\code{\link{update_tbl}()}
}
\concept{tbls}
