% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{tbls}
\alias{tbls}
\title{List NocoDB tables}
\usage{
tbls(
  id_base = base_id(origin = origin, email = email, password = password, api_token =
    api_token),
  include_m2m = TRUE,
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token")
)
}
\arguments{
\item{id_base}{\code{\link[base:character]{character(1)}}\cr
NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}.}

\item{include_m2m}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to include many-to-many junction tables in the result.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} with metadata about tables in the specified base on a NocoDB server from its
\href{https://meta-apis-v2.nocodb.com/#tag/DB-Table/operation/db-table-list}{\code{GET /api/v2/meta/bases/{id_base}/tables}} API endpoint.
}
\seealso{
Other functions to manage NocoDB tables:
\code{\link{create_data_src_tbl}()},
\code{\link{create_tbl}()},
\code{\link{data_src_tbls}()},
\code{\link{delete_tbl}()},
\code{\link{reorder_tbl}()},
\code{\link{set_tbl_metadata}()},
\code{\link{tbl}},
\code{\link{tbl_id}()},
\code{\link{update_tbl}()}
}
\concept{tbls}
