% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{test_data_src}
\alias{test_data_src}
\title{Test NocoDB data source}
\usage{
test_data_src(
  connection,
  type = c("mssql", "mysql", "pg", "sqlite3"),
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token"),
  quiet = FALSE
)
}
\arguments{
\item{connection}{Type-specific connection details for the data source. A list.}

\item{type}{Type of the data source. One of \code{"mssql"}, \code{"mysql"}, \code{"pg"} or \code{"sqlite3"}.}

\item{origin}{NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
A logical scalar indicating whether the connection test was successful or not.
}
\description{
Tests a data source's connection details via a NocoDB server's
\href{https://meta-apis-v2.nocodb.com/#tag/Utils/operation/utils-test-connection}{\code{POST /api/v2/meta/connection/test}} API endpoint.
}
\details{
Note that the PostgreSQL \code{connection} parameters \code{sslmode} and \code{database} have no effect, i.e. are simply ignored by the API endpoint.
}
\examples{
\dontrun{
nocodb::test_data_src(type = "pg",
                      connection = list(host = "REPLACE-ME",
                                        port = 5432L,
                                        ssl = TRUE,
                                        user = "REPLACE-ME",
                                        password = "REPLACE-ME"))}
}
\seealso{
Other functions to manage NocoDB base data sources:
\code{\link{create_data_src}()},
\code{\link{create_data_src_tbl}()},
\code{\link{data_src}()},
\code{\link{data_src_diff}()},
\code{\link{data_src_id}()},
\code{\link{data_src_tbls}()},
\code{\link{data_srcs}()},
\code{\link{delete_data_src}()},
\code{\link{has_data_src_diff}()},
\code{\link{sync_data_src}()},
\code{\link{update_data_src}()}
}
\concept{data_src}
