% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{test_plugin}
\alias{test_plugin}
\title{Test NocoDB plugin configuration}
\usage{
test_plugin(
  title,
  config,
  category = plugin_category(id_plugin = plugin_id(title = title)),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg, required = FALSE),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg, required = FALSE),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg, required = FALSE)
)
}
\arguments{
\item{title}{NocoDB plugin title. A character scalar.}

\item{config}{Plugin configuration. A list.}

\item{category}{NocoDB plugin category as returned by \code{\link[=plugin_category]{plugin_category()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
\code{TRUE} if the test was successful.
}
\description{
Tests the configuration of the plugin with the specified title and category on a NocoDB server via its
\href{https://docs.nocodb.com/0.109.7/developer-resources/rest-apis/#meta-apis}{\code{POST /api/v1/db/meta/plugins/test}} API endpoint.
}
\seealso{
Other functions to manage NocoDB plugins:
\code{\link{is_plugin_active}()},
\code{\link{plugin}()},
\code{\link{plugin_category}()},
\code{\link{plugin_id}()},
\code{\link{plugins}()},
\code{\link{update_plugin}()}
}
\concept{plugins}
