% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{update_base}
\alias{update_base}
\title{Update NocoDB base}
\usage{
update_base(
  title = NULL,
  description = NULL,
  color = NULL,
  show_null_and_empty_in_filter = NULL,
  id_base = base_id(origin = origin, email = email, password = password, api_token =
    api_token),
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token")
)
}
\arguments{
\item{title}{\code{\link[base:character]{character(1)}}\cr
New base title.}

\item{description}{\code{\link[base:character]{character(1)}}\cr
Base description.}

\item{color}{\code{\link[base:character]{character(1)}}\cr
Color to use for the base icon. A hexadecimal color code like "#36BFFF".}

\item{show_null_and_empty_in_filter}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to distinguish \code{NULL}s from empty values in column filters. If \code{FALSE}, \code{NULL} and empty fields are
treated alike.}

\item{id_base}{\code{\link[base:character]{character(1)}}\cr
NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}
}
\value{
\code{id_base}, invisibly.
}
\description{
Updates the specified base on a NocoDB server via its \href{https://meta-apis-v2.nocodb.com/#tag/Base/operation/base-update}{\code{PATCH /api/v2/meta/bases/{id_base}}}
API endpoint.
}
\seealso{
Other functions to manage NocoDB bases:
\code{\link{base}()},
\code{\link{base_id}()},
\code{\link{base_ui_acl}()},
\code{\link{base_users}()},
\code{\link{bases}()},
\code{\link{create_base}()},
\code{\link{delete_base}()},
\code{\link{delete_base_user}()},
\code{\link{invite_base_user}()},
\code{\link{resend_base_user_invitation}()},
\code{\link{update_base_ui_acl}()},
\code{\link{update_base_user}()}
}
\concept{bases}
