% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{update_base_user}
\alias{update_base_user}
\title{Update NocoDB base user}
\usage{
update_base_user(
  user_email,
  role = c("no-access", "viewer", "commenter", "editor", "creator"),
  id_user = user_id(user_email = user_email, id_base = id_base, hostname = hostname,
    email = email, password = password, api_token = api_token),
  id_base = base_id(hostname = hostname, email = email, password = password, api_token =
    api_token),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg),
  quiet = FALSE
)
}
\arguments{
\item{user_email}{E-mail address of the user. A character scalar.}

\item{role}{\href{https://docs.nocodb.com/collaboration/oss-specific-details/#base-level-permissions}{Base role} to assign to the user. One of
\code{"no-access"}, \code{"viewer"}, \code{"commenter"}, \code{"editor"} or \code{"creator"}.}

\item{id_user}{NocoDB user identifier as returned by \code{\link[=user_id]{user_id()}}. A character scalar.}

\item{id_base}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{id_user}, invisibly.
}
\description{
Updates the specified user in regard to the specified base on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/Auth/operation/auth-base-user-update}{\code{PATCH /api/v2/meta/bases/{id_base}/users/{id_user}}} API endpoint.
}
\details{
Only the super admin user is allowed to use this API endpoint, i.e. the provided credentials must belong to them.
}
\seealso{
Other functions to manage NocoDB users:
\code{\link{add_user}()},
\code{\link{base_users}()},
\code{\link{delete_base_user}()},
\code{\link{delete_user}()},
\code{\link{invite_base_user}()},
\code{\link{invite_user}()},
\code{\link{resend_base_user_invitation}()},
\code{\link{sign_up_user}()},
\code{\link{update_user}()},
\code{\link{user_id}()},
\code{\link{users}()},
\code{\link{validate_user_email}()},
\code{\link{whoami}()}

Other functions to manage NocoDB bases:
\code{\link{base}()},
\code{\link{base_id}()},
\code{\link{base_users}()},
\code{\link{bases}()},
\code{\link{create_base}()},
\code{\link{delete_base}()},
\code{\link{delete_base_user}()},
\code{\link{invite_base_user}()},
\code{\link{resend_base_user_invitation}()},
\code{\link{update_base}()}
}
\concept{bases}
\concept{users}
