% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{update_data_src}
\alias{update_data_src}
\title{Update NocoDB data source}
\usage{
update_data_src(
  id_data_src,
  connection = NULL,
  type = NULL,
  alias = NULL,
  inflection_column = NULL,
  inflection_table = NULL,
  enabled = NULL,
  id_base = base_id(hostname = hostname, email = email, password = password, api_token =
    api_token),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)
}
\arguments{
\item{id_data_src}{NocoDB data source identifier as returned by \code{\link[=data_src_id]{data_src_id()}}. A character scalar.}

\item{connection}{Type-specific connection details for the data source. A list.}

\item{type}{Type of the data source. One of \code{"mssql"}, \code{"mysql"}, \code{"pg"} or \code{"sqlite3"}.}

\item{alias}{Alias of the data source. A character scalar.}

\item{inflection_column}{Type of inflection to apply for column names in the data source. One of
\code{"none"} or \code{"camelize"}.}

\item{inflection_table}{Type of inflection to apply for table names in the data source. One of
\code{"none"} or \code{"camelize"}.}

\item{enabled}{Whether the added data source is enabled or disabled.}

\item{id_base}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with metadata about the updated data source.
}
\description{
Updates the specified data source of the specified base on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/Source/operation/source-update}{\code{PATCH /api/v2/meta/bases/{id_base}/sources/{id_data_src}}} API endpoint.
}
\seealso{
Other functions to manage NocoDB base data sources:
\code{\link{create_data_src}()},
\code{\link{data_src}()},
\code{\link{data_src_id}()},
\code{\link{data_srcs}()},
\code{\link{delete_data_src}()},
\code{\link{test_data_src}()}
}
\concept{data_src}
