% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{update_tbl}
\alias{update_tbl}
\title{Update NocoDB table metadata}
\usage{
update_tbl(
  tbl_id,
  body_json,
  auto_unbox = TRUE,
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  auth_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg),
  quiet = FALSE
)
}
\arguments{
\item{tbl_id}{NocoDB table identifier as returned by \code{\link[=tbl_id]{tbl_id()}}. A character scalar.}

\item{body_json}{Data to include as JSON in the HTTP request body. Set to \code{NULL} for an empty body.}

\item{auto_unbox}{Whether or not to automatically "unbox" length-1 vectors in \code{body_json} to JSON scalars.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{auth_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API authentication token}. A character scalar.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{tbl_id}, invisibly.
}
\description{
Updates the metadata of the specified table on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/DB-Table/operation/db-table-update}{\code{PATCH /api/v2/meta/tables/{tbl_id}}} API endpoint.
}
\seealso{
Other functions to manage NocoDB tables:
\code{\link{reorder_tbl}()},
\code{\link{set_tbl_metadata}()},
\code{\link{tbl}},
\code{\link{tbl_id}()},
\code{\link{tbls}()}
}
\concept{tbls}
