% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{update_tbl}
\alias{update_tbl}
\title{Update NocoDB table}
\usage{
update_tbl(
  id_tbl,
  id_base = NULL,
  name = NULL,
  title = NULL,
  meta = NULL,
  hostname = pal::pkg_config_val("hostname"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token"),
  quiet = FALSE
)
}
\arguments{
\item{id_tbl}{NocoDB table identifier as returned by \code{\link[=tbl_id]{tbl_id()}}. A character scalar.}

\item{id_base}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{name}{Table name. A character scalar.}

\item{title}{NocoDB-specific table title. A character scalar.}

\item{meta}{NocoDB-specific table metadata. A list.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{id_tbl}, invisibly.
}
\description{
Updates the metadata of the specified table on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/DB-Table/operation/db-table-update}{\code{PATCH /api/v2/meta/tables/{id_tbl}}} API endpoint.
}
\seealso{
Other functions to manage NocoDB tables:
\code{\link{create_data_src_tbl}()},
\code{\link{create_tbl}()},
\code{\link{data_src_tbls}()},
\code{\link{delete_tbl}()},
\code{\link{reorder_tbl}()},
\code{\link{set_tbl_metadata}()},
\code{\link{tbl}},
\code{\link{tbl_id}()},
\code{\link{tbls}()}
}
\concept{tbls}
