% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{update_tbl_col}
\alias{update_tbl_col}
\title{Update NocoDB table column}
\usage{
update_tbl_col(
  id_col,
  column_name = NULL,
  title = NULL,
  description = NULL,
  uidt = NULL,
  dt = NULL,
  cdf = NULL,
  origin = pal::pkg_config_val("origin"),
  email = pal::pkg_config_val("email"),
  password = pal::pkg_config_val("password"),
  api_token = pal::pkg_config_val("api_token")
)
}
\arguments{
\item{id_col}{NocoDB column identifier as returned by \code{\link[=tbl_col_id]{tbl_col_id()}}. A character scalar.}

\item{column_name}{Column name. Either \code{NULL} to omit or a character scalar.}

\item{title}{NocoDB column title. Either \code{NULL} to omit or a character scalar.}

\item{description}{NocoDB column description displayed as a tooltip in the column header. Either \code{NULL} to omit or a character scalar.}

\item{uidt}{NocoDB \strong{u}ser \strong{i}nterface \strong{d}ata \strong{t}ype. Either \code{NULL} to omit or one of
\itemize{
\item \code{"Attachment"}
\item \code{"AutoNumber"}
\item \code{"Barcode"}
\item \code{"Button"}
\item \code{"Checkbox"}
\item \code{"Collaborator"}
\item \code{"Count"}
\item \code{"CreatedBy"}
\item \code{"CreatedTime"}
\item \code{"Currency"}
\item \code{"Date"}
\item \code{"DateTime"}
\item \code{"Decimal"}
\item \code{"Duration"}
\item \code{"Email"}
\item \code{"ForeignKey"}
\item \code{"Formula"}
\item \code{"GeoData"}
\item \code{"Geometry"}
\item \code{"ID"}
\item \code{"JSON"}
\item \code{"LastModifiedBy"}
\item \code{"LastModifiedTime"}
\item \code{"Links"}
\item \code{"LinkToAnotherRecord"}
\item \code{"LongText"}
\item \code{"Lookup"}
\item \code{"MultiSelect"}
\item \code{"Number"}
\item \code{"Percent"}
\item \code{"PhoneNumber"}
\item \code{"QrCode"}
\item \code{"Rating"}
\item \code{"Rollup"}
\item \code{"SingleLineText"}
\item \code{"SingleSelect"}
\item \code{"SpecificDBType"}
\item \code{"Time"}
\item \code{"URL"}
\item \code{"User"}
\item \code{"Year"} .
}}

\item{dt}{Column data type. Either \code{NULL} to omit or a character scalar.}

\item{cdf}{Column default value. Either \code{NULL} to omit or a character scalar.}

\item{origin}{NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with metadata about the NocoDB table to which the updated column belongs, invisibly.
}
\description{
Updates the metadata of the specified table column on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/DB-Table-Column/operation/db-table-column-update}{\code{PATCH /api/v2/meta/columns/{id_col}}} API endpoint.

Beware that this API endpoint alters the schema of the underlying table, which might be undesirable, especially if the table is from an external \href{https://docs.nocodb.com/category/data-sources}{data source}.
}
\seealso{
Other functions to manage NocoDB table columns:
\code{\link{create_tbl_col}()},
\code{\link{delete_tbl_col}()},
\code{\link{set_display_val}()},
\code{\link{set_display_vals}()},
\code{\link{tbl_col}()},
\code{\link{tbl_col_id}()},
\code{\link{tbl_cols}()}
}
\concept{cols}
