% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{update_tbl_col}
\alias{update_tbl_col}
\title{Update NocoDB table column}
\usage{
update_tbl_col(
  id_col,
  column_name = NULL,
  title = NULL,
  description = NULL,
  uidt = NULL,
  colOptions = NULL,
  dt = NULL,
  cdf = NULL,
  rqd = NULL,
  pk = NULL,
  ai = NULL,
  au = NULL,
  un = NULL,
  dtx = NULL,
  dtxp = NULL,
  dtxs = NULL,
  enable_rich_text = NULL,
  origin = funky::config_val("origin"),
  email = funky::config_val("email"),
  password = funky::config_val("password"),
  api_token = funky::config_val("api_token"),
  quiet = FALSE
)
}
\arguments{
\item{id_col}{\code{\link[base:character]{character(1)}}\cr
NocoDB column identifier as returned by \code{\link[=tbl_col_id]{tbl_col_id()}}.}

\item{column_name}{\code{\link[base:character]{character(1)}}\cr
Column name. Omitted if \code{NULL}.}

\item{title}{\code{\link[base:character]{character(1)}}\cr
NocoDB column title. Omitted if \code{NULL}.}

\item{description}{\code{\link[base:character]{character(1)}}\cr
NocoDB column description displayed as a tooltip in the column header. Omitted if \code{NULL}. Set \code{description = ""} to remove an existing \code{description} value.}

\item{uidt}{\code{\link[base:character]{character(1)}}\cr
NocoDB \strong{u}ser \strong{i}nterface \strong{d}ata \strong{t}ype. Either \code{NULL} for the column's current \code{uidt} value or one of
\itemize{
\item \code{"Attachment"}
\item \code{"AutoNumber"}
\item \code{"Barcode"}
\item \code{"Button"}
\item \code{"Checkbox"}
\item \code{"Collaborator"}
\item \code{"Count"}
\item \code{"CreatedBy"}
\item \code{"CreatedTime"}
\item \code{"Currency"}
\item \code{"Date"}
\item \code{"DateTime"}
\item \code{"Decimal"}
\item \code{"Duration"}
\item \code{"Email"}
\item \code{"ForeignKey"}
\item \code{"Formula"}
\item \code{"GeoData"}
\item \code{"Geometry"}
\item \code{"ID"}
\item \code{"JSON"}
\item \code{"LastModifiedBy"}
\item \code{"LastModifiedTime"}
\item \code{"Links"}
\item \code{"LinkToAnotherRecord"}
\item \code{"LongText"}
\item \code{"Lookup"}
\item \code{"MultiSelect"}
\item \code{"Number"}
\item \code{"Percent"}
\item \code{"PhoneNumber"}
\item \code{"QrCode"}
\item \code{"Rating"}
\item \code{"Rollup"}
\item \code{"SingleLineText"}
\item \code{"SingleSelect"}
\item \code{"SpecificDBType"}
\item \code{"Time"}
\item \code{"URL"}
\item \code{"User"}
\item \code{"Year"} .
}}

\item{colOptions}{\code{\link[base:list]{list()}}\cr
NocoDB column options. To specify the values for a single-select field (\code{uidt = "SingleSelect"}), supply a data frame with the columns \code{title}, \code{color}
(HEX color code) and \code{index} (counted from 0 upwards) as a list element named \code{options}. Defaults to the column's current \code{colOptions} value if \code{NULL}.}

\item{dt}{\code{\link[base:character]{character(1)}}\cr
Column \strong{d}ata \strong{t}ype. Omitted if \code{NULL}.}

\item{cdf}{\code{\link[base:character]{character(1)}}\cr
\strong{C}olumn \strong{d}efault \strong{v}alue. Defaults to the column's current \code{cdf} value if \code{NULL}. Set \code{cdf = NA} to remove an existing \code{cdf} value.}

\item{rqd}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not the column is \strong{r}e\strong{q}uire\strong{d} to be filled, i.e. should have a \verb{NOT NULL} constraint. Corresponds to the \code{NN} \href{https://docs.nocodb.com/fields/field-types/custom-types/specific-db-type/#advanced-field-attributes}{advanced field attribute}. Defaults to the column's current \code{rqd}
value if \code{NULL}.}

\item{pk}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not the column is a \strong{p}rimary \strong{k}ey, should have a \verb{PRIMARY KEY} constraint. Corresponds to the \code{PK} \href{https://docs.nocodb.com/fields/field-types/custom-types/specific-db-type/#advanced-field-attributes}{advanced field attribute}. Defaults to the column's current \code{pk}
value if \code{NULL}.}

\item{ai}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to \strong{a}uto-\strong{i}ncrement the column value on new record insertion. Corresponds to the \code{AI} \href{https://docs.nocodb.com/fields/field-types/custom-types/specific-db-type/#advanced-field-attributes}{advanced field attribute}. Defaults to the column's current \code{ai}
value if \code{NULL}.}

\item{au}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to \strong{a}uto-\strong{u}pdate the column value to the current timestamp on record creation/insertion. Corresponds to the \code{AU} \href{https://docs.nocodb.com/fields/field-types/custom-types/specific-db-type/#advanced-field-attributes}{advanced field attribute}. Defaults to the column's current \code{au}
value if \code{NULL}.}

\item{un}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not the column is an \strong{un}signed number, i.e. can only store non-negative numbers. Corresponds to the \code{UN} \href{https://docs.nocodb.com/fields/field-types/custom-types/specific-db-type/#advanced-field-attributes}{advanced field attribute}. Defaults to the column's current \code{un}
value if \code{NULL}.}

\item{dtx}{\strong{D}ata \strong{t}ype \strong{x}. Defaults to the column's current \code{dtx} value if \code{NULL}.}

\item{dtxp}{\strong{D}ata \strong{t}ype \strong{x} \strong{p}recision. Omitted if \code{NULL}.}

\item{dtxs}{\strong{D}ata \strong{t}ype \strong{x} \strong{s}cale. Omitted if \code{NULL}.}

\item{enable_rich_text}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to enable \href{https://docs.nocodb.com/fields/field-types/text-based/rich-text/}{rich text} support for the column in NocoDB. Only applicable
if \code{uidt = "LongText"}. Omitted if \code{NULL}.}

\item{origin}{\code{\link[base:character]{character(1)}}\cr
NocoDB server \href{https://developer.mozilla.org/docs/Glossary/Origin}{origin}.}

\item{email}{\code{\link[base:character]{character(1)}}\cr
E-mail address of the NocoDB user to authenticate with.}

\item{password}{\code{\link[base:character]{character(1)}}\cr
Password of the NocoDB user to authenticate with.}

\item{api_token}{\code{\link[base:character]{character(1)}}\cr
NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if provided.}

\item{quiet}{\code{\link[base:logical]{logical(1)}}\cr
Whether or not to suppress printing status output from internal processing.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with metadata about the NocoDB table to which the updated column belongs, invisibly.
}
\description{
Updates the metadata of the specified table column on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/Fields/operation/db-table-column-update}{\code{PATCH /api/v2/meta/columns/{id_col}}} API endpoint.

Beware that this API endpoint \emph{always} alters the schema of the underlying table, which might be undesirable, especially if the table belongs to an external
\href{https://docs.nocodb.com/category/data-sources}{data source}.
}
\seealso{
Other functions to manage NocoDB table columns:
\code{\link{create_tbl_col}()},
\code{\link{delete_tbl_col}()},
\code{\link{set_display_val}()},
\code{\link{set_display_vals}()},
\code{\link{tbl_col}()},
\code{\link{tbl_col_id}()},
\code{\link{tbl_cols}()}
}
\concept{cols}
