% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{update_tbl_col}
\alias{update_tbl_col}
\title{Update NocoDB table column metadata}
\usage{
update_tbl_col(
  col_id,
  body_json,
  auto_unbox = TRUE,
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  auth_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg)
)
}
\arguments{
\item{col_id}{NocoDB column identifier as returned by \code{\link[=tbl_col_id]{tbl_col_id()}}. A character scalar.}

\item{body_json}{Data to include as JSON in the HTTP request body. Set to \code{NULL} for an empty body.}

\item{auto_unbox}{Whether or not to automatically "unbox" length-1 vectors in \code{body_json} to JSON scalars.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{auth_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API authentication token}. A character scalar.}
}
\value{
TODO
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Updates the metadata of the specified table column on a NocoDB server via its
\href{https://meta-apis-v2.nocodb.com/#tag/DB-Table-Column/operation/db-table-column-update}{\code{PATCH /api/v2/meta/columns/{col_id}}} API endpoint.
}
\seealso{
Other functions to manage NocoDB table columns:
\code{\link{set_display_val}()},
\code{\link{set_display_vals}()},
\code{\link{tbl_col}()},
\code{\link{tbl_col_id}()},
\code{\link{tbl_cols}()}
}
\concept{cols}
