% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nocodb.gen.R
\name{user_id}
\alias{user_id}
\title{Get NocoDB user ID}
\usage{
user_id(
  user_email,
  id_base = base_id(),
  hostname = pal::pkg_config_val(key = "hostname", pkg = this_pkg),
  email = pal::pkg_config_val(key = "email", pkg = this_pkg, required = FALSE),
  password = pal::pkg_config_val(key = "password", pkg = this_pkg, required = FALSE),
  api_token = pal::pkg_config_val(key = "api_token", pkg = this_pkg, required = FALSE)
)
}
\arguments{
\item{user_email}{E-mail address of the user. A character scalar.}

\item{id_base}{NocoDB base identifier as returned by \code{\link[=base_id]{base_id()}}. A character scalar.}

\item{hostname}{NocoDB server \href{https://en.wikipedia.org/wiki/Hostname}{hostname}. A character scalar.}

\item{email}{E-mail address of the NocoDB user to authenticate with.}

\item{password}{Password of the NocoDB user to authenticate with.}

\item{api_token}{NocoDB \href{https://docs.nocodb.com/account-settings/api-tokens/}{API token}. Takes precedence over \code{email} and \code{password} if
provided.}
}
\value{
User identifier as a character scalar.
}
\description{
Returns the identifier of the user with the specified display \code{name} on a NocoDB server.
}
\seealso{
Other functions to manage NocoDB users:
\code{\link{base_users}()},
\code{\link{invite_user}()},
\code{\link{sign_up_user}()},
\code{\link{update_base_user}()},
\code{\link{update_user}()},
\code{\link{user}()},
\code{\link{validate_user_email}()}
}
\concept{users}
